@@ -7,6 +7,8 @@ import {
   StyleSheet,
   ScrollView,
   Clipboard,
+  StatusBar,
+  Platform
 } from 'react-native'
 
 import { pubS,DetailNavigatorStyle } from '../../styles/'
@@ -15,14 +17,10 @@ import { TextInputComponent,Btn } from '../../components/'
 import Modal from 'react-native-modal'
 import QRCode from 'react-native-qrcode'
 import { connect } from 'react-redux'
-import I18n from 'react-native-i18n'
-import UserSQLite from '../../utils/accountDB'
-import TradingSQLite from '../../utils/tradingDB'
-
-const sqLite = new UserSQLite()
-let db
+import I18n, { getLanguages } from 'react-native-i18n'
+import { changeBackupModalTimesAction } from '../../actions/accountManageAction'
 const wallet = require('ethereumjs-wallet')
-import Toast from 'react-native-toast'
+import Toast from 'react-native-root-toast'
 class Receive extends Component{
   constructor(props){
     super(props)
@@ -30,63 +28,125 @@ class Receive extends Component{
         visible: false,
         payTotalVal: '',
         addressText: '',
-        currentAccount: { }
+
+        backupBtnStyle_height: scaleSize(70),
+        backupBtnStyle_width: scaleSize(500),
+
+        whileView_height: scaleSize(420),
+
+        blueView_height: scaleSize(355),
+        blueView_padding: scaleSize(30),
+
+        modalView_width: scaleSize(620),
+        modalView_height: scaleSize(865),
+        modalView_top: scaleSize(49),
     }
 
   }
 
-  componentWillMount(){
+  componentDidMount(){
 
-    if(!db){  
-        db = sqLite.open();  
-    }
+    const { currentAccount,backupModalTimes } = this.props.accountManageReducer
 
-    db.transaction((tx) => {
-        tx.executeSql("select * from account ", [], (tx,results) => {
-
-          let len = results.rows.length 
-          
-          for(let i=0; i<len; i++){  
-            let result = results.rows.item(i)
-            if(result.is_selected === 1){
-              this.setState({
-                currentAccount: result,
-                addressText: `0x${result.address}`
-              })
-            }
-            
-            if(result.backup_status === 0){
-              this.setState({
-                visible: true
-              })
-            }
-          } 
-        },(error) => {
-          console.error(error)
-        })
+    this.setState({
+      addressText: `0x${currentAccount.address}`
+    })
+    if(currentAccount.backup_status === 0 && backupModalTimes === 0){
+      this.setState({
+        visible: true
       })
+    }
+    // console.log('I18n.currentLoca=====',I18n.currentLocale())
+    let a  =I18n.currentLocale()
 
+    I18n.currentLocale().search(/zh/) === 0 ? a = 'zh' : null
+    
+    if(Platform.OS === 'ios'){
+      switch(a){
+        case 'zh':
+          this.onZh()
+          break
+        case 'en':
+          this.onEn()
+          break
+        case 'ru-RU':
+          this.onRu()
+          break
+        default:
+          break
+      }
+    }else{
+      switch(I18n.currentLocale()){
+        case 'zh-CN':
+          this.onZh()
+          break
+        case 'en-US':
+          this.onEn()
+          break
+        case 'ru-RU':
+          this.onRu()
+          break
+        default:
+          break
+      }
+    }
+    this.props.dispatch(changeBackupModalTimesAction(1))
   }
 
   componentWillUnmount(){
-    // this.getContent()
     this.setState({
       visible: false
     }) 
   }
-  // async getContent(){
-  //   let a= await Clipboard.getString()
-  // }
+
 
   // onChangePayTotal = (val) => {
   //   this.setState({
   //     payTotalVal: val,
   //   })
   // }
-
+  onZh = () => {
+    this.setState({
+      backupBtnStyle_height: scaleSize(70),
+      backupBtnStyle_width: scaleSize(500),
+      whileView_height: scaleSize(356) ,
+      blueView_height: scaleSize(260),
+      modalView_width: scaleSize(560),
+      modalView_height: scaleSize(700),
+      modalView_top: scaleSize(149),
+      blueView_padding: scaleSize(60),
+    })
+  }
+  onEn = () => {
+    this.setState({
+      backupBtnStyle_height: scaleSize(70),
+      backupBtnStyle_width: scaleSize(500),
+      whileView_height:scaleSize(356) ,
+      blueView_height: scaleSize(360),
+      modalView_width: scaleSize(560),
+      modalView_height: scaleSize(800),
+      modalView_top: scaleSize(120),
+      blueView_padding: scaleSize(30),
+    })
+  }
+  onRu = () => {
+    this.setState({
+      backupBtnStyle_height: scaleSize(70),
+      backupBtnStyle_width: scaleSize(500),
+      whileView_height:scaleSize(456) ,
+      blueView_height: scaleSize(420),
+      modalView_width: scaleSize(660),
+      modalView_height: scaleSize(960),
+      modalView_top: scaleSize(80),
+      blueView_padding: scaleSize(0),
+    })
+  }
   onPressCopyBtn = () => {
       Clipboard.setString(this.state.addressText)
-      Toast.showLongBottom(I18n.t('copy_successfully'))
+      let t = Toast.show(I18n.t('copy_successfully'))
+      setTimeout(() => {
+        Toast.hide(t)
+      },1000)
   }
 
   onHide = () => {
@@ -95,25 +155,31 @@ class Receive extends Component{
     })
   }
   backupBtn = () => {
-    const { currentAccount } = this.state
-    console.log('currentAccount===',currentAccount)
+    const { currentAccount,globalAccountsList } = this.props.accountManageReducer
+
     this.setState({
       visible: false
     })
+    
     this.props.navigator.push({
       screen: 'back_up_account',
       title: currentAccount.account_name,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       navigatorStyle: DetailNavigatorStyle,
       passProps: {
         userName: currentAccount.account_name,
         address: currentAccount.address,
         b_id: currentAccount.id,
+        accountsNumber: globalAccountsList.length,
+        currentAccountId: currentAccount.id,
+        psdPrompt: currentAccount.password_promp || ''
       },
       // navigatorButtons: {
       //   rightButtons: [
       //     {
       //       title: 'Save',
-      //       id: 'save_back_up_info'
+      //       id: 'save_change'
       //     }
       //   ]
       // }
@@ -121,9 +187,15 @@ class Receive extends Component{
   }
 
   render(){
-    const { payTotalVal,visible,addressText } = this.state
+    const { payTotalVal,visible,addressText, backupBtnStyle_height,backupBtnStyle_width, whileView_height, blueView_height, modalView_width, modalView_height,modalView_top,blueView_padding } = this.state
     return(
       <View style={[pubS.container,{paddingTop: scaleSize(35)}]}>
+      {
+          Platform.OS === 'ios' ?
+          <StatusBar backgroundColor="#000000"  barStyle="dark-content" animated={true} />
+          : null
+      }
+      
         {
         // <TextInputComponent
         //   placeholder={'Enter receive amount (Optional)'}
@@ -153,30 +225,38 @@ class Receive extends Component{
         isVisible={visible}
         onBackButtonPress={this.onHide}
         onBackdropPress={this.onHide}
-        style={styles.modalView}
+        style={[styles.modalView,{height: modalView_height+ scaleSize(75),width: modalView_width,top: modalView_top}]}
         backdropOpacity={.8}
       >
-        <Image source={require('../../images/xhdpi/img_collectionnobackup.png')} style={styles.modalImageStyle}/>
-        <View style={[{alignItems:'center'}]}>
-            <View style={styles.blueView}>
-                <Text style={[pubS.font36_3,{marginTop: scaleSize(32)}]}>{I18n.t('backup_first')}</Text>
-                <Text style={[pubS.font22_2,{marginTop: scaleSize(13),width: '90%',textAlign:'center'}]}>{I18n.t('backup_modal_1')}</Text>
-            </View>
-            <View style={styles.whileView}>
-                <Text style={[pubS.font30_2,{marginTop: scaleSize(25)}]}>{`--  ${I18n.t('backup_mnemonic')} --`}</Text>
-                <Text style={[pubS.font24_2,{textAlign:'center'}]}>{I18n.t('backup_modal_2')}</Text>
-                <Text style={[pubS.font30_2,{marginTop: scaleSize(25)}]}>{`--  ${I18n.t('backup_keystore')}  --`}</Text>
-                <Text style={[pubS.font24_2,{textAlign:'center'}]}>{I18n.t('backup_modal_3')}</Text>
-                <TouchableOpacity activeOpacity={.7} onPress={this.backupBtn} style={[styles.backupBtnStyle,pubS.center]}>
-                  <Text style={pubS.font28_2}>{I18n.t('backup_now')}</Text>
-                </TouchableOpacity>
-            </View>
-      </View>
+        <Image source={require('../../images/xhdpi/img_collectionnobackup.png')} style={[styles.modalImageStyle,{ left: (modalView_width-scaleSize(107))/2,}]}/>
+        <View style={[{alignItems:'center',}]}>
+          <View style={[styles.blueView,{height: blueView_height,paddingLeft: blueView_padding,paddingRight: blueView_padding}]}>
+              <Text style={[pubS.font36_3,{marginTop: scaleSize(32),textAlign:'center'}]}>{I18n.t('backup_first')}</Text>
+              <Text style={[pubS.font22_2,{marginTop: scaleSize(13),width: '100%',textAlign:'center'}]}>{I18n.t('backup_modal_1')}</Text>
+          </View>
+          <View style={[styles.whileView,{height: whileView_height}]}>
+              <Text style={[pubS.font30_2,{marginTop: scaleSize(20),textAlign:'center'}]}>{`--  ${I18n.t('backup_mnemonic')} --`}</Text>
+              <Text style={[pubS.font24_2,{textAlign:'center'}]}>{I18n.t('backup_modal_2')}</Text>
+              <Text style={[pubS.font30_2,{marginTop: scaleSize(20),textAlign:'center'}]}>{`--  ${I18n.t('backup_keystore')}  --`}</Text>
+              <Text style={[pubS.font24_2,{textAlign:'center'}]}>{I18n.t('backup_modal_3')}</Text>
+              <TouchableOpacity 
+                activeOpacity={.7} 
+                onPress={this.backupBtn} 
+                style={[styles.backupBtnStyle,pubS.center,{height: backupBtnStyle_height,width:backupBtnStyle_width}]}
+              >
+                <Text style={pubS.font28_2}>{I18n.t('backup_now')}</Text>
+              </TouchableOpacity>
+          </View>
+        </View>
+        <TouchableOpacity activeOpacity={.7} onPress={this.onHide} style={[styles.closeViewStyle,pubS.center]}>
+          <Image source={require('../../images/xhdpi/btn_ico_collectionnobackup_close_def.png')} style={{height: scaleSize(44),height: scaleSize(44)}}/>
+        </TouchableOpacity>
       </Modal>
       </View>
     )
   }
 }
+
 const styles = StyleSheet.create({
   modalImageStyle:{
     height: scaleSize(121),
@@ -184,43 +264,45 @@ const styles = StyleSheet.create({
     zIndex:999,
     position:'absolute',
     top:0,
-    left: scaleSize(226.5),
+  },
+  closeViewStyle: {
+    height: scaleSize(75),
+    width: scaleSize(75),
+    zIndex:999,
+    position:'absolute',
+    bottom:0,
+    alignSelf:'center'
+    // borderColor:'red',
+    // borderWidth:1,
   },
   backupBtnStyle:{
-    height: scaleSize(70),
-    width: scaleSize(500),
+      
     borderWidth: 1,
     borderColor: '#2B8AFF',
     borderRadius: scaleSize(35),
     marginTop: scaleSize(30),
   },
   whileView:{
-      height: scaleSize(420),
-      backgroundColor:'#fff',
-      width:'100%',
-      alignItems:'center',
-      borderBottomLeftRadius :scaleSize(10),
-      borderBottomRightRadius :scaleSize(10),
+    backgroundColor:'#fff',
+    width:'100%',
+    alignItems:'center',
+    borderBottomLeftRadius :scaleSize(10),
+    borderBottomRightRadius :scaleSize(10),
   },
   blueView: {
-    height: scaleSize(355),
     backgroundColor:'#2B8AFF',
     width:'100%',
     alignItems:'center',
     borderTopLeftRadius: scaleSize(10),
     borderTopRightRadius: scaleSize(10),
-    marginTop: scaleSize(84)
+    // marginTop: scaleSize(84),
   },
   modalView:{
-    width: scaleSize(620),
-    height: scaleSize(865),
     position: 'absolute',
-    top: scaleSize(59),
     alignSelf: 'center',
     backgroundColor:'transparent',
-    // borderColor:'#fff',
+    // borderColor:'red',
     // borderWidth:1,
-
   }
 })
 export default connect(
