@@ -17,11 +17,9 @@ import com.reactnativenavigation.NavigationApplication;
 
 import com.bitgo.randombytes.RandomBytesPackage;
 import org.reactnative.camera.RNCameraPackage;
-import io.realm.react.RealmReactPackage;
 import org.pgsqlite.SQLitePluginPackage;
-import com.chirag.RNMail.*;
 import com.AlexanderZaytsev.RNI18n.RNI18nPackage;
-import com.remobile.toast.RCTToastPackage; 
+import com.learnium.RNDeviceInfo.RNDeviceInfo;
 public class MainApplication extends NavigationApplication {
 
     @Override
@@ -34,11 +32,9 @@ public class MainApplication extends NavigationApplication {
           new PickerViewPackage(),
           new RandomBytesPackage(),
           new RNCameraPackage(),
-          new RealmReactPackage(),
           new SQLitePluginPackage(),
-          new RNMail(),
           new RNI18nPackage(),
-          new RCTToastPackage()
+          new RNDeviceInfo()
         );
     }
 
