@@ -1,90 +1,136 @@
 import * as types from '../constants/tokenManageConstant'
 import tokenDBOpation from '../utils/tokenDBOpation'
-const insertToTokenAction = (addr) => {
-	const onInsert = () => {
+
+const gloablTokenList = (list) => {
+	const getList = () => {
 		return {
-			type: types.INSERT_TO_TOKEN_DB,
-			payload: {	
-				currentAddress: addr
+			type: types.GLOBAL_TOKEN_LIST,
+			payload: {
+				list
 			}
 		}
 	}
-	return(dispatch,getState) => {
-		dispatch(onInsert())
+	return (dispatch,getState) => {
+		dispatch(getList())
 	}
 }
-const getAssetsListAction = () => {
-	const onGet = () => {
+const fetchTokenAction = (addr,refresh) => {
+	const onFetchStart = () => {
 		return {
-			type: types.GET_ASSETS_LIST,
+			type: types.FETCH_TOKEN_LIST,
+			payload: {
+				addr
+			}
 		}
 	}
-	return(dispatch,getState) => {
-		dispatch(onGet())
+	const suc = (list) => {
+		return {
+			type: types.FETCH_TOKEN_LIST_SUC,
+			payload: {
+				list
+			}	
+		}
+	}
+	const err = () => {
+		return {
+			type: types.FETCH_TOKEN_LIST_ERR,
+			payload: {
+
+			}	
+		}
+	}
+	return (dispatch,getState) => {
+		dispatch(onFetchStart())
+		tokenDBOpation.fetchToken({
+			parames: {
+				addr,
+				refresh
+			},
+			fetchTokenSuccess: (list) => {dispatch(suc(list))},
+			fetchTokenFail: (msg) => {dispatch(err(msg))}
+		})
 	}
 }
-// const selectedTokenListAction = (sele) => {
-// 	console.log('发出的action',sele)
-// 	const list = () => {
-// 		return {
-// 			type: types.SELECTED_TOKEN_LIST,
-// 			payload:{
-// 				sele
-// 			}
-// 		}
-// 	}
-// 	return(dispatch,getState) => {
-// 		dispatch(list())
-// 	}
-// }
 
-const deleteSelectedToListAction = (delAddr,asList) => {
+
+const deleteSelectedToListAction = (delAddr,curaddr) => {
+	const onDeleteStart = () => {
+		return {
+			type: types.DELETE_TOKEN_LIST_START,
+			payload:{
+				delAddr,
+				curaddr,
+			}
+		}
+	}
+
 	const onDelete = () => {
 		return {
 			type: types.DELETE_TOKEN_LIST,
-			payload:{
-				delAddr,
-				asList
-			}
+			
 		}
 	}
+	
 	return(dispatch,getState) => {
-		dispatch(onDelete())
+		dispatch(onDeleteStart())
+		tokenDBOpation.deleteSelectedToken({
+			parames: {
+				delAddr,
+				curaddr
+			},
+			deleteSelected: (data) => {dispatch(onDelete(data))}
+		})
 	}
 }
 
-const addSelectedToListAction = (addAddr,asList) => {
+const addSelectedToListAction = (addAddr,curaddr) => {
+	const onAddStart = () => {
+		return {
+			type: types.ADD_TOKEN_LIST_START,
+			payload:{
+				addAddr,
+				curaddr
+			}
+		}
+	}
 	const onAdd = () => {
 		return {
 			type: types.ADD_TOKEN_LIST,
-			payload:{
-				addAddr,
-				asList
-			}
+			
 		}
 	}
 	return(dispatch,getState) => {
-		dispatch(onAdd())
+		dispatch(onAddStart())
+		tokenDBOpation.addSelectedToken({
+			parames: {
+				addAddr,
+				curaddr
+			},
+			addSelected: (data) => {dispatch(onAdd(data))}
+		})
 	}
 }
-const initSelectedListAction = (data,addr) => {
-	// console.log('initSelectedListAction',data)
-	// console.log('initSelectedListAction',addr)
-	const onInit = () => {
+const initSelectedListAction = () => {
+	const onInitSel = (data) => {
 		return {
 			type: types.INIT_SELECTED_LIST,
 			payload: {
-				initList: data,
-				curAddr: addr
+				data
 			}
 		}
 	}
 	return(dispatch,getState) => {
-		dispatch(onInit())
+		tokenDBOpation.initSelectedToken({
+			parames: {
+
+			},
+			initSelectedTokenList: (data) => {dispatch(onInitSel(data))}
+		})
 	}
 }
 
-const refreshTokenInfoAction = (addr) => {
+const refreshTokenAction = (addr,tokenlist) => {
+	console.log('下拉刷新action')
 	const onRef = () => {
 		return {
 			type: types.REFRESH_TOKEN_INFO,
@@ -97,25 +143,73 @@ const refreshTokenInfoAction = (addr) => {
 		return {
 			type: types.REFRESH_TOKEN_SUCCESS,
 			payload: {
+				data,
+				
+			}
+		}
+	}
+	const fail = (msg) => {
+		return {
+			type: types.REFRESH_TOKEN_FAIL,
+			payload: {
+				msg
+			}
+		}
+	}
+	const etz = (data) => {
+		return {
+			type: types.REFRESH_ETZ,
+			payload: {
 				data
 			}
 		}
 	}
 	return(dispatch,getState) => {
 		dispatch(onRef())
-		tokenDBOpation.refresh({
+		tokenDBOpation.tokenRefresh({
 			parames:{
 				addr,
+				tokenlist
 			},
-			refreshSuccess: (data) => { dispatch(suc(data))}
+			refreshEtz: (data) => { dispatch(etz(data))},
+			refreshSuccess: (data) => { dispatch(suc(data))},
+			refreshFail: (msg) => { dispatch(fail(msg))}
 		})
 	}
 }
+
+const switchTokenAction = (addr) => {
+	const onSwitchStart = () => {
+		return {
+			type:types.SWITCH_TOKEN_LIST_START,
+		}
+	}
+	const onSwitch = (data) => {
+		return {
+			type: types.SWITCH_TOKEN_LIST,
+			payload: {
+				data
+			}
+		}
+	}
+	return(dispatch,getState) => {
+		dispatch(onSwitchStart())
+		tokenDBOpation.switchTokenList({
+			parames: {
+				addr
+			},
+			switchTokenSuc: (data) => { dispatch(onSwitch(data))}
+		})
+	}
+}
+
+
 export {
-	insertToTokenAction,
-	getAssetsListAction,
 	deleteSelectedToListAction,
 	addSelectedToListAction,
 	initSelectedListAction,
-	refreshTokenInfoAction,
+	refreshTokenAction,
+	fetchTokenAction,
+	gloablTokenList,
+	switchTokenAction,
 }
