@@ -1,10 +1,14 @@
 import TextInputComponent from './TextInputComponent'
 import Btn from './Btn'
 import Loading from './Loading'
+import LoadingModal from './LoadingModal'
 import ArrowToDetail from './ArrowToDetail'
+import NavHeader from './NavHeader'
 export {
   TextInputComponent,
   Btn,
   Loading,
-  ArrowToDetail
+  LoadingModal,
+  ArrowToDetail,
+  NavHeader
 }
