@@ -1,6 +1,6 @@
-import { StyleSheet, View, Text, Image } from 'react-native'
+import { BackHandler } from 'react-native'
 import { Navigation, ScreenVisibilityListener } from 'react-native-navigation'
-
+import { onExitApp } from '../utils/exitApp'
 import store from '../store/'
 import { Provider } from 'react-redux'
 
@@ -15,6 +15,7 @@ import CreateAccountSuccess from '../views/login/CreateAccountSuccess'
 
 import Personal from '../views/personal/'
 import Support from '../views/personal/Support'
+import ModifyPassword from '../views/personal/ModifyPassword'
 import HelpCenter from '../views/personal/HelpCenter'
 import AccountManage from '../views/personal/AccountManage'
 import Setting from '../views/personal/setting/'
@@ -22,7 +23,7 @@ import SwitchLanguage from '../views/personal/setting/SwitchLanguage'
 import BackUpAccount from '../views/personal/backup/'
 import WriteMnemonic from '../views/personal/backup/WriteMnemonic'
 import VerifyMnemonic from '../views/personal/backup/VerifyMnemonic'
-
+import ExportKeyStore from '../views/personal/backup/ExportKeyStore'
 
 import Assets from '../views/home/Assets'
 import MsgCenterList from '../views/home/MsgCenterList'
@@ -30,13 +31,18 @@ import Receive from '../views/home/Receive'
 import Payment from '../views/home/Payment'
 import TradingRecord from '../views/home/tradingRecord/'
 import ScanQrCode from '../views/home/ScanQrCode'
-import AssetDetailList from '../views/home/AssetDetailList'
+import TxRecordlList from '../views/home/TxRecordlList'
 import TradingRecordDetail from '../views/home/TradingRecordDetail'
 import TxWebView from '../views/home/TxWebView'
 import SwitchWallet from '../views/home/SwitchWallet'
 import AddAssets from '../views/home/AddAssets'
+import BindPhone from '../views/home/BindPhone'
+import RecomPrize from '../views/home/RecomPrize'
+import SelectCoutry from '../views/home/SelectCoutry'
+import DownLoadApp from '../views/home/DownLoadApp'
+import ReceiveCandy from '../views/home/ReceiveCandy'
 
-export function registerScreens() {
+function registerScreens() {
   Navigation.registerComponent('splash', () => Splash,store,Provider)
   Navigation.registerComponent('login', () => Login,store,Provider)
   Navigation.registerComponent('create_account', () => CreateAccount,store,Provider)
@@ -53,7 +59,7 @@ export function registerScreens() {
   Navigation.registerComponent('on_receive', () => Receive,store,Provider)
   Navigation.registerComponent('trading_record', () => TradingRecord,store,Provider)
   Navigation.registerComponent('scan_qr_code', () => ScanQrCode,store,Provider)
-  Navigation.registerComponent('asset_detail_list', () => AssetDetailList,store,Provider)
+  Navigation.registerComponent('tx_record_list', () => TxRecordlList,store,Provider)
   Navigation.registerComponent('trading_record_detail', () => TradingRecordDetail,store,Provider)
   Navigation.registerComponent('support', () => Support,store,Provider)
   Navigation.registerComponent('help_center', () => HelpCenter,store,Provider)
@@ -63,15 +69,42 @@ export function registerScreens() {
   Navigation.registerComponent('tx_web_view', () => TxWebView,store,Provider)
   Navigation.registerComponent('switch_wallet', () => SwitchWallet,store,Provider)
   Navigation.registerComponent('add_assets', () => AddAssets,store,Provider)
-
+  Navigation.registerComponent('bind_phone', () => BindPhone,store,Provider)
+  Navigation.registerComponent('recom_prize', () => RecomPrize,store,Provider)
+  Navigation.registerComponent('select_coutry', () => SelectCoutry,store,Provider)
+  Navigation.registerComponent('download_app', () => DownLoadApp,store,Provider)
+  Navigation.registerComponent('receive_candy', () => ReceiveCandy,store,Provider)
+  Navigation.registerComponent('modify_password', () => ModifyPassword,store,Provider)
+  Navigation.registerComponent('export_keystore', () => ExportKeyStore,store,Provider)
 }
 
 // 注册页面切换监听器
-export function registerScreenVisibilityListener() {
+function registerScreenVisibilityListener() {
   new ScreenVisibilityListener({
-    // willAppear: ({screen}) => console.log(`Displaying screen ${screen}`),
+    // willAppear: ({screen}) => addFunc(screen),
     // didAppear: ({screen, startTime, endTime, commandType}) => console.log('screenVisibility', `Screen ${screen} displayed in ${endTime - startTime} millis [${commandType}]`),
-    // willDisappear: ({screen}) => console.log(`Screen will disappear ${screen}`),
+    // willDisappear: ({screen}) => removeFunc(screen),
     // didDisappear: ({screen}) => console.log(`Screen disappeared ${screen}`)
   }).register()
+}
+
+// function addFunc(screen){
+//   if(screen === 'home_assets'){
+//     BackHandler.addEventListener('hardwareBackPress',() => {
+//       onExitApp()
+//     })
+//   }
+// }
+// function removeFunc(screen){
+//   if(screen === 'home_assets'){
+//     BackHandler.removeEventListener('hardwareBackPress',() => {
+//       onExitApp()
+//     })
+//   }
+// }
+
+
+export {
+  registerScreens,
+  registerScreenVisibilityListener
 }
