@@ -1,18 +1,13 @@
-const GET_ACCOUNT_INFO = 'GET_ACCOUNT_INFO'
-const ON_SWITCH_ACCOUNT = 'ON_SWITCH_ACCOUNT'
+const ON_SWITCH_ACCOUNT_START = 'ON_SWITCH_ACCOUNT_START'
+const ON_SWITCH_ACCOUNT_END = 'ON_SWITCH_ACCOUNT_END'
 const ON_IMPORT_ACCOUNT = 'ON_IMPORT_ACCOUNT'
-// const IMPORT_ACCOUNT_SUCC = 'IMPORT_ACCOUNT_SUCC'
-// const IMPORT_ACCOUNT_FAIL = 'IMPORT_ACCOUNT_FAIL'
-// const ON_DELETE_ACCOUNT = 'ON_DELETE_ACCOUNT'
+
 const RESET_DELETE_STATUS = 'RESET_DELETE_STATUS'
 const UPDATE_BACKUP_STATUS = 'UPDATE_BACKUP_STATUS'
 const ON_DELETE_ACCOUNT_START = 'ON_DELETE_ACCOUNT_START'
 const ON_DELETE_ACCOUNT_SUC = 'ON_DELETE_ACCOUNT_SUC'
 const ON_DELETE_ACCOUNT_FAIL = 'ON_DELETE_ACCOUNT_FAIL'
 
-const PASS_ACCOUNTS_INFO_START = 'PASS_ACCOUNTS_INFO_START'
-const PASS_ACCOUNTS_INFO_SUC = 'PASS_ACCOUNTS_INFO_SUC'
-const PASS_ACCOUNTS_INFO_FAIL = 'PASS_ACCOUNTS_INFO_FAIL'
 
 const CREATE_ACCOUNT_START = 'CREATE_ACCOUNT_START'
 const CREATE_ACCOUNT_SUC = 'CREATE_ACCOUNT_SUC'
@@ -24,23 +19,44 @@ const IMPORT_ACCOUNT_SUC = 'IMPORT_ACCOUNT_SUC'
 const IMPORT_ACCOUNT_FAIL = 'IMPORT_ACCOUNT_FAIL'
 
 const DELETE_MNEMONIC = 'DELETE_MNEMONIC'
+const DELETE_MNEMONIC_START = 'DELETE_MNEMONIC_START'
+
+const GLOBAL_ALL_ACCOUNTS_INFO = 'GLOBAL_ALL_ACCOUNTS_INFO'
+const GLOBAL_CURRENT_ACCOUNT_INFO = 'GLOBAL_CURRENT_ACCOUNT_INFO'
+
+const UPDATE_BACKUP_STATUS_SUCC = 'UPDATE_BACKUP_STATUS_SUCC'
+
+const CHANGE_BACKUP_MODAL_TIMES = 'CHANGE_BACKUP_MODAL_TIMES'
+
+const SHOW_IMPORT_LOADING = 'SHOW_IMPORT_LOADING'
+
+const PASS_SCAN_RECEIVE_ADDRESS = 'PASS_SCAN_RECEIVE_ADDRESS'
+
+const REFERSH_MANEGE_BALANCE = 'REFERSH_MANEGE_BALANCE'
+
+const MODIFY_PASSWORD_SUC = 'MODIFY_PASSWORD_SUC'
+const MODIFY_PASSWORD_FAIL = 'MODIFY_PASSWORD_FAIL'
+const MODIFY_PASSWORD_START = 'MODIFY_PASSWORD_START'
+const MODIFY = 'MODIFY'
+const PASS_PROPS = 'PASS_PROPS'
+
+
+const GEN_MNEMONIC_START = 'GEN_MNEMONIC_START'
+const GEN_MNEMONIC_SUC = 'GEN_MNEMONIC_SUC'
+
+const COPY_KEYSTORE_BACKUP = 'COPY_KEYSTORE_BACKUP'
 export {
-	GET_ACCOUNT_INFO,
-	ON_SWITCH_ACCOUNT,
+	ON_SWITCH_ACCOUNT_START,
+	ON_SWITCH_ACCOUNT_END,
 	ON_IMPORT_ACCOUNT,
-	// ON_DELETE_ACCOUNT,
+
 	RESET_DELETE_STATUS,
 	UPDATE_BACKUP_STATUS,	
-	// IMPORT_ACCOUNT_SUCC,
-	// IMPORT_ACCOUNT_FAIL,
+
 
 	ON_DELETE_ACCOUNT_START,
 	ON_DELETE_ACCOUNT_SUC,
 	ON_DELETE_ACCOUNT_FAIL,
-
-	PASS_ACCOUNTS_INFO_START,
-	PASS_ACCOUNTS_INFO_SUC,
-	PASS_ACCOUNTS_INFO_FAIL,
 
 	CREATE_ACCOUNT_START,
 	CREATE_ACCOUNT_SUC,
@@ -50,5 +66,31 @@ export {
 	IMPORT_ACCOUNT_SUC,
 	IMPORT_ACCOUNT_FAIL,
 
-	DELETE_MNEMONIC
+	DELETE_MNEMONIC,
+	DELETE_MNEMONIC_START,
+
+	GLOBAL_ALL_ACCOUNTS_INFO,
+
+	GLOBAL_CURRENT_ACCOUNT_INFO,
+
+	UPDATE_BACKUP_STATUS_SUCC,
+
+	CHANGE_BACKUP_MODAL_TIMES,
+
+	SHOW_IMPORT_LOADING,
+
+	PASS_SCAN_RECEIVE_ADDRESS,
+
+	REFERSH_MANEGE_BALANCE,
+
+	MODIFY_PASSWORD_SUC,
+	MODIFY_PASSWORD_FAIL,
+	MODIFY_PASSWORD_START,
+	MODIFY,
+	PASS_PROPS,
+
+	GEN_MNEMONIC_START,
+	GEN_MNEMONIC_SUC,
+
+	COPY_KEYSTORE_BACKUP
 }
