@@ -1,18 +1,52 @@
 const INSERT_TO_TOKEN_DB = 'INSERT_TO_TOKEN_DB'
-const GET_ASSETS_LIST = 'GET_ASSETS_LIST'
+
 const SELECTED_TOKEN_LIST = 'SELECTED_TOKEN_LIST'
+
+const DELETE_TOKEN_LIST_START = 'DELETE_TOKEN_LIST_START'
 const DELETE_TOKEN_LIST = 'DELETE_TOKEN_LIST'
+
+const ADD_TOKEN_LIST_START = 'ADD_TOKEN_LIST_START'
 const ADD_TOKEN_LIST = 'ADD_TOKEN_LIST'
+
 const INIT_SELECTED_LIST = 'INIT_SELECTED_LIST'
 const REFRESH_TOKEN_INFO = 'REFRESH_TOKEN_INFO'
 const REFRESH_TOKEN_SUCCESS = 'REFRESH_TOKEN_SUCCESS'
+const REFRESH_TOKEN_FAIL = 'REFRESH_TOKEN_FAIL'
+
+const FETCH_TOKEN_LIST = 'FETCH_TOKEN_LIST'
+const FETCH_TOKEN_LIST_SUC = 'FETCH_TOKEN_LIST_SUC'
+const FETCH_TOKEN_LIST_ERR = 'FETCH_TOKEN_LIST_ERR'
+
+const GLOBAL_TOKEN_LIST = 'GLOBAL_TOKEN_LIST'
+
+const SWITCH_TOKEN_LIST = 'SWITCH_TOKEN_LIST'
+const SWITCH_TOKEN_LIST_START = 'SWITCH_TOKEN_LIST_START'
+
+const REFRESH_ETZ = 'REFRESH_ETZ'
+
 export {
 	INSERT_TO_TOKEN_DB,
-	GET_ASSETS_LIST,
 	SELECTED_TOKEN_LIST,
+	
+	DELETE_TOKEN_LIST_START,
 	DELETE_TOKEN_LIST,
+
+	ADD_TOKEN_LIST_START,
 	ADD_TOKEN_LIST,
+	
 	INIT_SELECTED_LIST,
 	REFRESH_TOKEN_INFO,
-	REFRESH_TOKEN_SUCCESS
+	REFRESH_TOKEN_SUCCESS,
+	REFRESH_TOKEN_FAIL,
+	
+	FETCH_TOKEN_LIST,
+	FETCH_TOKEN_LIST_SUC,
+	FETCH_TOKEN_LIST_ERR,
+
+	GLOBAL_TOKEN_LIST,
+
+	SWITCH_TOKEN_LIST,
+	SWITCH_TOKEN_LIST_START,
+
+	REFRESH_ETZ
 }
