@@ -1,4 +1,43 @@
-const INSERT_TO_DB = 'INSERT_TO_DB'
+const SAVE_TO_RECORD_START = 'SAVE_TO_RECORD_START'
+const SAVE_TO_RECORD_SUC = 'SAVE_TO_RECORD_SUC'
+const SAVE_TO_RECORD_FAIL = 'SAVE_TO_RECORD_FAIL'
+
+const MAKE_TX_BY_ETZ_START = 'MAKE_TX_BY_ETZ_START'
+const MAKE_TX_BY_ETZ_SUC = 'MAKE_TX_BY_ETZ_SUC'
+const MAKE_TX_BY_ETZ_FAIL = 'MAKE_TX_BY_ETZ_FAIL'
+
+const MAKE_TX_BY_TOEKN_START = 'MAKE_TX_BY_TOEKN_START'
+const MAKE_TX_BY_TOKEN_SUC = 'MAKE_TX_BY_TOKEN_SUC'
+const MAKE_TX_BY_TOKEN_FAIL = 'MAKE_TX_BY_TOKEN_FAIL'
+
+
+const MAKE_TX_BY_ETZ_HASH = 'MAKE_TX_BY_ETZ_HASH'
+const RESET_TX_STATUS = 'RESET_TX_STATUS'
+
+
+const SHOW_LOADING = 'SHOW_LOADING'
+const UPDATE_TX_LIST = 'UPDATE_TX_LIST'
+
+const TX_ETZ_PSD_ERROR = 'TX_ETZ_PSD_ERROR'
+
 export {
-	INSERT_TO_DB,
+	SAVE_TO_RECORD_START,
+	SAVE_TO_RECORD_SUC,
+	SAVE_TO_RECORD_FAIL,
+
+	MAKE_TX_BY_ETZ_START,
+	MAKE_TX_BY_ETZ_SUC,
+	MAKE_TX_BY_ETZ_FAIL,
+	MAKE_TX_BY_ETZ_HASH,
+	RESET_TX_STATUS,
+
+	MAKE_TX_BY_TOEKN_START,
+	MAKE_TX_BY_TOKEN_SUC,
+	MAKE_TX_BY_TOKEN_FAIL,
+
+	SHOW_LOADING,
+	UPDATE_TX_LIST,
+
+	TX_ETZ_PSD_ERROR
+
 }
