@@ -3,15 +3,13 @@ const wallet = require('ethereumjs-wallet')
 const hdkey = require('ethereumjs-wallet/hdkey')
 const util = require('ethereumjs-util')
 const bip39 = require('bip39')
-import UserSQLite from '../utils/accountDB'
 
-const sqLite = new UserSQLite()
-let db
 
 const initState = {
 	accountInfo: [],
 	currentAddr: '',
 	importStatus: '',
+	importFailMsg: '',
 	// deleteFinished: false,
 	updateBackupSucc: false,
 	isLoading: false,
@@ -19,14 +17,39 @@ const initState = {
 	passAccInfoSuc: '',
 	createSucc: false,
 	delMnemonicSuc: false,
+
+	globalAccountsList: [],
+	currentAccount: {},
+	deleteCurrentAccount: false,
+	backupModalTimes: 0,
+	importLoading: false,
+
+	scanAddress: '',
+	scanCurToken: '',
+	allAmoountAddress: [],
+
+	modifyStatus: 0,
+	modifyText: '',
+
+	pass_currentList: {},
+	pass_keyStore: {},
+
+	mnemonicValue: '',
+	create_usernane: '',
+	create_psd: '',
+	create_prompt: '',
+	create_from: '',
+
+	copyKeystoreSuc: false,
 }
 export default function accountManageReducer (state = initState,action) {
 	switch(action.type){
-		case types.GET_ACCOUNT_INFO:
-			return getAccInfo(state,action)
+
+		case types.ON_SWITCH_ACCOUNT_START:
+			return onSwitchStart(state,action)
 			break
-		case types.ON_SWITCH_ACCOUNT:
-			return onSwitch(state,action)
+		case types.ON_SWITCH_ACCOUNT_END:
+			return onSwitchEnd(state,action)
 			break
 		case types.ON_DELETE_ACCOUNT_START:
 			return onDelStart(state,action)
@@ -41,25 +64,16 @@ export default function accountManageReducer (state = initState,action) {
 			return onReset(state,action)
 			break
 		case types.UPDATE_BACKUP_STATUS:
-			return onUpateBackup(state,action)
+			return onUpateBackupStart(state,action)
 			break
-		case types.PASS_ACCOUNTS_INFO_START:
-			return passInfoStart(state,action)
-			break
-		case types.PASS_ACCOUNTS_INFO_SUC:
-			return passInfoSuc(state,action)
-			break
-		case types.PASS_ACCOUNTS_INFO_FAIL:
-			return passInfoFail(state,action)
+		case types.UPDATE_BACKUP_STATUS_SUCC:
+			return onUpateBackupSuc(state,action)
 			break
 		case types.CREATE_ACCOUNT_START:
-			return createStart(state,action)
+			return accountCreateStart(state,action)
 			break
 		case types.CREATE_ACCOUNT_SUC:
-			return createSuc(state,action)
-			break
-		case types.CREATE_ACCOUNT_FAIL:
-			return createFail(state,action)
+			return accountCreateSuc(state,action)
 			break
 		case types.IMPORT_ACCOUNT_START:
 			return importStart(state,action)
@@ -73,31 +87,198 @@ export default function accountManageReducer (state = initState,action) {
 		case types.DELETE_MNEMONIC:
 			return onDelMneMonic(state,action)
 			break
+		case types.DELETE_MNEMONIC_START:
+			return onDelMneMonicStart(state,action)
+			break
+		case types.GLOBAL_ALL_ACCOUNTS_INFO:
+			return globalAccounts(state,action)
+			break
+		case types.GLOBAL_CURRENT_ACCOUNT_INFO:
+			return globalCurrentAccounts(state,action)
+			break
+		case types.CHANGE_BACKUP_MODAL_TIMES:
+			return changeBackupTimes(state,action)
+			break
+		case types.SHOW_IMPORT_LOADING:
+			return onShowLoading(state,action)
+			break
+		case types.PASS_SCAN_RECEIVE_ADDRESS:
+			return onPassScanAddr(state,action)
+			break
+		case types.REFERSH_MANEGE_BALANCE:
+			return onBalance(state,action)
+			break
+		case types.MODIFY_PASSWORD_SUC:
+			return onModifySuc(state,action)
+			break
+		case types.MODIFY_PASSWORD_FAIL:
+			return onModifyFail(state,action)
+			break	
+		case types.MODIFY_PASSWORD_START:
+			return onModifyStart(state,action)
+			break	
+		case types.PASS_PROPS:
+			return onPassProps(state,action)
+			break
+		case types.GEN_MNEMONIC_START:
+			return genMneStart(state,action)
+			break
+		case types.GEN_MNEMONIC_SUC:
+			return genMneSuc(state,action)
+			break
+		case types.COPY_KEYSTORE_BACKUP:
+			return copyKeyBackup(state,action)
+			break
 		default:
 			return state
 			break
 
 	}
 }
-
-const onDelMneMonic = (state,action) => {
-	const { addr } = action.payload
-	if(!db){  
-	      db = sqLite.open();  
-	} 
-    db.transaction((tx)=>{  
-      	tx.executeSql("update account set mnemonic = '' where address= ? ", [addr],(tx,results)=>{  
-			
-       	})  
-      	},(error)=>{
-        console.error(error)
-    }) 
-	return {
+const copyKeyBackup = (state,action) => {
+	return{
 		...state,
-		delMnemonicSuc: true
+		copyKeystoreSuc: true
 	}
 }
-const createStart = (state,action) => {
+const genMneStart = (state,action) => {
+	return {
+		...state,
+		mnemonicValue: '',
+		create_usernane: '',
+		create_psd: '',
+		create_prompt: '',
+		create_from: ''
+	}
+}
+const genMneSuc = (state,action) => {
+	const { mne,userNameVal, psdVal, promptVal,fromLogin }  = action.payload
+
+	return {
+		...state,
+		mnemonicValue: mne,
+		create_usernane: userNameVal,
+		create_psd: psdVal,
+		create_prompt: promptVal,
+		create_from: fromLogin
+	}
+}
+const onPassProps = (state,action) => {
+	const { currentList, keyStore} = action.payload	
+	return {
+		...state,
+		pass_currentList: currentList,
+		pass_keyStore: keyStore
+	}
+}
+
+const onModifyStart = (state,action) => {
+
+	console.log('777777777777globalAccountsList',state.globalAccountsList)
+	return {
+		...state,
+		modifyStatus: 0,
+		modifyText: '',
+	}
+}
+
+const onModifySuc = (state,action) => {
+	const { modifyText,modifyResult } = action.payload
+	console.log('modifyResult===',modifyResult[0])
+	const { kid, ciphertext, mac, salt, iv, address} = modifyResult[0]
+
+	let newState = Object.assign({},state)
+
+	console.log('newState.globalAccountsList===',newState.globalAccountsList)
+	newState.globalAccountsList.map((value,index) => {
+		if(value.address === address){
+
+			 newState.globalAccountsList[index].ciphertext = ciphertext
+
+			 newState.globalAccountsList[index].mac = mac
+
+			 newState.globalAccountsList[index].salt = salt
+
+			 newState.globalAccountsList[index].iv = iv
+
+
+			 newState.globalAccountsList[index].kid = kid
+		}
+	})
+
+	return {
+		...newState,
+		modifyStatus: 1,
+		modifyText: modifyText,
+	}
+}
+
+const onModifyFail = (state,action) => {
+	const { msg } = action.payload
+	return {
+		...state,
+		modifyStatus: 2,
+		modifyText:msg
+	}
+}
+const onPassScanAddr = (state,action) => {
+	const { addr, token} = action.payload
+	return {
+		...state,
+		scanAddress: addr,
+		scanCurToken: token,
+	}
+}
+
+const changeBackupTimes = (state,action) => {
+	const { time } = action.payload
+	return {
+		...state,
+		backupModalTimes: time,
+	}
+}
+
+const onShowLoading = (state,action) => {
+	const { status } = action.payload
+	return{
+		...state,
+		importLoading: status
+	}
+}
+const globalCurrentAccounts = (state,action) => {
+	return {
+		...state,
+		currentAccount: action.payload.currinfos,
+	}
+}
+
+const globalAccounts = (state,action) => {
+	return {
+		...state,
+		globalAccountsList: action.payload.infos
+	}
+}
+const onDelMneMonicStart = (state,action) => {
+	return {
+		...state,
+		delMnemonicSuc: false,
+	}
+}
+const onDelMneMonic = (state,action) => {
+	const { data,addr } = action.payload
+
+	let newState = Object.assign({},state)
+	newState.globalAccountsList.map((list,index) => {
+		if(list.address === addr){
+			newState.globalAccountsList[index].mnemonic = ''
+		}
+	})
+	return {
+		...state,
+		delMnemonicSuc: data
+	}
+}
+const accountCreateStart = (state,action) => {
 	console.log('创建开始',)
 	return {
 		...state,
@@ -105,42 +286,38 @@ const createStart = (state,action) => {
 		createSucc: false
 	}
 }
-const createSuc = (state,action) => {
+const accountCreateSuc = (state,action) => {
+	const { data } = action.payload
 	console.log('创建完成')
 	return {
 		...state,
 		isLoading: false,
-		createSucc: true
+		createSucc: data
 	}
 }
-const createFail = (state,action) => {
-	return {
-		...state,
-		isLoading: false,
-		createSucc: false
-	}
-}
-const onUpateBackup = (state,action) => {
-	const { addr } = action.payload
-	let status = false
-	if(!db){  
-       db = sqLite.open();  
-    } 
-	db.transaction((tx)=>{  
-      tx.executeSql("update account set backup_status = 1 where address= ? ", [addr],(tx,results)=>{  
-        console.log('更新成功')  
-        // status = true
-      }) 
-    },(error)=>{
-      console.log('更新失败',error)
-    })
 
+const onUpateBackupStart = (state,action) => {
 
+	return state
+}
+
+const onUpateBackupSuc = (state,action) => {
+	const { data,updateAddr } = action.payload
+
+	let newState = Object.assign({},state)
+
+	newState.globalAccountsList.map((list,index) => {
+		if(list.address === updateAddr){
+			// list.backup_status = 1
+			newState.globalAccountsList[index].backup_status = 1
+		}
+	})
 	return {
-		...state,
-		updateBackupSucc: true
+		...newState,
+		updateBackupSucc: data
 	}
 }
+
 const onReset = (state,action) => {
 	return {
 		...state,
@@ -149,6 +326,7 @@ const onReset = (state,action) => {
 		deleteSuc: false,
 		createSucc: false,
 		delMnemonicSuc: false,
+		deleteCurrentAccount: false
 	}
 }
 
@@ -159,16 +337,40 @@ const onDelStart = (state,action) => {
 	}
 }
 const onDelSuc = (state,action) => {
-	const { data } = action.payload
+	const { deleteId,curId } = action.payload
+
+	let newState = Object.assign({},state)
+
+	newState.globalAccountsList.map((list,index) => {
+		
+		if(list.id === deleteId){
+			newState.globalAccountsList.splice(index,1)
+		}	
+
+	})
+	//删除的是当前账号  is_selected = 1
+	console.log('deleteId==',deleteId)
+	console.log('curId==',curId)
+	console.log('newState.globalAccountsList===',newState.globalAccountsList)
+	let delCur = false 
+	if(deleteId === curId){
+		if(newState.globalAccountsList.length > 0){
+			newState.globalAccountsList[0].is_selected = 1
+			delCur = true
+		}
+	}
+
+
+	console.log('删除成功',newState.globalAccountsList)
 
 	return {
-		...state,
+		...newState,
 		isLoading: false,
-		deleteSuc: true
+		deleteSuc: true,
+		deleteCurrentAccount: delCur
 	}
 }
 const onDelFail = (state,action) => {
-	const { msg } = action.payload
 	console.log('删除失败')
 	return {
 		...state,
@@ -177,66 +379,36 @@ const onDelFail = (state,action) => {
 	}
 }
 	
-const getAccInfo = (state,action) => {
-	return {
-		...state,
-		accountInfo: action.payload.info
-	}
+
+const onSwitchStart = (state,action) => {
+	
+
+
+	return state
 }
 
-const passInfoStart = (state,action) => {
-	return {
-		...state,
-	}
-}
-const passInfoSuc = (state,action) => {
-	const { data } = action.payload
-	console.log('data passInfoSuc====',data)
-	return {
-		...state,
-		accountInfo: data,
-		passAccInfoSuc: 'home',
-	}
-}
-const passInfoFail = (state,action) => {
-	return {
-		...state,
-		passAccInfoSuc: 'login'
-	}
-}
+const onSwitchEnd = (state,action) => {
+	const { switchAddr } = action.payload
 
-const onSwitch = (state,action) => {
-	const addr = action.payload.addr
-	const newState = Object.assign({},state)
-	let curr =  ''
+	let newState = Object.assign({},state)
 
-	newState.accountInfo.map((val,index) => {
-		if(addr === val.address){
-			curr = val.address
-
-			if(!db){  
-		        db = sqLite.open();  
-		    }  
-		    db.transaction((tx)=>{  
-		      tx.executeSql("update account set is_selected = 1 where address= ? ", [curr],(tx,results)=>{  
-		        db.transaction((tx) => {
-		        	tx.executeSql("update account set is_selected = 0 where address != ? ", [curr],(tx,results)=>{
-		        	})
-		        }, (error) => {
-		        	console.error(error)
-		        })
-		      });  
-		    },(error)=>{
-		      console.error(error)
-		    })
-		}	
+	newState.globalAccountsList.map((list,index) => {
+		console.log('list.address==',list.address)
+		console.log('switchAddr==',switchAddr)
+		if(list.address === switchAddr){
+			newState.globalAccountsList[index].is_selected = 1
+			newState.currentAccount = newState.globalAccountsList[index]
+		}else{
+			newState.globalAccountsList[index].is_selected = 0
+		}
 	})
+
+
+	console.log('newState.globalAccountsList===',newState.globalAccountsList)
 	return {
-		...state,
-		currentAddr: curr,
+		...newState,
 	}
 }
-
 
 const importStart = (state,action) => {
 	return {
@@ -245,14 +417,35 @@ const importStart = (state,action) => {
 	}
 }
 const importSuc = (state,action) => {
+	const { data } = action.payload //导入的这条数据
+	
+	let newState = Object.assign({},state)
+
+	newState.globalAccountsList.push(data)
+
 	return {
-		...state,
+		...newState,
 		importStatus: 'success',
+		importLoading: false
 	}
 }
 const importFail = (state,action) => {
+	const { msg } = action.payload
 	return {
 		...state,
 		importStatus: 'fail',
+		importFailMsg: msg,
+		importLoading: false
 	}
 }
+const onBalance = (state,action) => {
+	const { data } = action.payload
+
+
+	let newState = Object.assign({},state)
+	for(let i = 0;i < data.balData.length; i ++){
+		newState.globalAccountsList[i].assets_total = data.balData[i]
+	}			
+	return newState
+}
+
