@@ -11,38 +11,53 @@ import {
   RefreshControl,
   Button,
   BackHandler,
+  StatusBar,
+  Alert,
+  AppState,
 } from 'react-native'
 
 import { pubS,DetailNavigatorStyle,MainThemeNavColor,ScanNavStyle } from '../../styles/'
-import { setScaleText, scaleSize } from '../../utils/adapter'
+import { setScaleText, scaleSize,ifIphoneX, } from '../../utils/adapter'
 import Drawer from 'react-native-drawer'
 import { connect } from 'react-redux'
-import { onSwitchDrawerAction } from '../../actions/onSwitchDrawerAction'
 import SwitchWallet from './SwitchWallet'
 import { switchDrawer } from '../../utils/switchDrawer'
-import TokenSQLite from '../../utils/tokenDB'
-import { toSplash } from '../../root'
-import { splitDecimal } from '../../utils/splitNumber'
-const tkSqLite = new TokenSQLite()
-let tk_db
-import { insertToTokenAction,initSelectedListAction,refreshTokenInfoAction } from '../../actions/tokenManageAction'
-let etzTitle = "ETZ"
+
+import { splitDecimal, scientificToNumber} from '../../utils/splitNumber'
+import {Scan} from '../../components/'
+
+import { insertToTokenAction,initSelectedListAction,refreshTokenAction,fetchTokenAction } from '../../actions/tokenManageAction'
+import { insert2TradingDBAction,resetTxStatusAction,updateTxListAction } from '../../actions/tradingManageAction'
 import I18n from 'react-native-i18n'
-import Toast from 'react-native-toast'
-// import { passAccountsInfoAction } from '../../actions/accountManageAction' 
+import Toast from 'react-native-root-toast'
+
+
+import { Navigation } from 'react-native-navigation'
+
+import accountDB from '../../db/account_db'
+
+import { globalAllAccountsInfoAction,globalCurrentAccountInfoAction } from '../../actions/accountManageAction'
+let etzTitle = "ETZ"
 
 class Assets extends Component{
   constructor(props){
     super(props)
     this.state = {
-      etzBalance: 0,
       navTitle: '',
       selectedAssetsList: [],
       isRefreshing: false,
       curAddr: '',
+      currencySymbol: '',
+      currentAppState: AppState.currentState,
     }
+    this.props.navigator.setOnNavigatorEvent(this.onNavigatorEvent.bind(this))
   }
+
   componentWillMount(){
+    // Navigation.dismissAllModals({
+    //   animationType: 'slide-down' // 'none' / 'slide-down' , dismiss animation for the modal (optional, default 'slide-down')
+    // })
+
     this.props.navigator.setTabButton({
       tabIndex: 0,
       label: I18n.t('assets')
@@ -51,96 +66,268 @@ class Assets extends Component{
       tabIndex:1,
       label: I18n.t('mine')
     })
-  }
-
-  componentDidMount(){
-    const { accountInfo } = this.props.accountManageReducer
     this.setState({
-      isRefreshing: true
+      isRefreshing: false
     })
-    accountInfo.map((val,index) => {
-      if(val.is_selected === 1){
-        this.onFetch(val.address)
-        
+    
+    localStorage.load({
+      key: 'lang',
+      autoSync: true,
+    }).then( ret => {
+      this.setCurrencySymbol(ret.selectedLan)
+    }).catch (err => {
+
+    })
+
+    
+  }
+  componentDidMount(){
+    AppState.addEventListener('change', this._handleAppStateChange)
+    this.getAllAccounts()
+  }
+  onNavigatorEvent(event) {
+    if (event.id === 'bottomTabSelected') {
+        this.onCloseDrawer()
+    }
+    if (event.id === 'bottomTabReselected') {
+      this.onCloseDrawer()
+    }
+
+    switch (event.id) {
+      case 'willAppear':
+        this.backHandler = BackHandler.addEventListener('hardwareBackPress', this.handleBackPress);
+        break;
+      case 'willDisappear':
+        this.backPressed = 0;
+        this.backHandler.remove();
+        break;
+      default:
+        break;
+    }
+
+  }
+   handleBackPress = () => {
+    if (this.backPressed && this.backPressed > 0) {
+      this.props.navigator.popToRoot({ animated: false });
+      return false;
+    }
+
+    this.backPressed = 1;
+
+    let t = Toast.show(I18n.t('click_again'))
+    setTimeout(() => {
+      Toast.hide(t)
+    },1000)
+
+    // this.props.navigator.showSnackbar({
+    //   text: 'Press one more time to exit',
+    //   duration: 'long',
+    // });
+    return true;
+  }
+
+  setCurrencySymbol(symbol){
+    switch(symbol){
+      case 'zh-CN':
         this.setState({
-          navTitle: val.account_name,
-          curAddr: val.address
+          currencySymbol: '¥',
         })
-        web3.eth.getBalance(`0x${val.address}`).then((res,rej)=>{
-          // console.log('res==',res)
-          this.setState({
-            etzBalance: web3.utils.fromWei(res,'ether')
-          })
+        break
+      case 'en-US':
+        this.setState({
+          currencySymbol: '$',
         })
+        break
+      case 'ru-RU':
+        this.setState({
+          currencySymbol: '₽',
+        })
+        break
+      default:
+        break
+    }
+  }
+  async getAllAccounts(){
+    const { fetchTokenList } = this.props.tokenManageReducer
+    //所有的账户数据
+    let findAccountsList = await accountDB.selectTable({
+      sql: 'select * from account',
+      parame: []
+    })
+    this.props.dispatch(globalAllAccountsInfoAction(findAccountsList))
+
+    findAccountsList.map((value,index) => {
+      if(value.is_selected === 1){
+        this.props.dispatch(refreshTokenAction(value.address,fetchTokenList))
+       
+        this.setState({
+          navTitle: value.account_name,
+          curAddr: value.address,
+          isRefreshing: true
+        })
+
+        
+        //当前账户信息
+        this.props.dispatch(globalCurrentAccountInfoAction(value))
+        //初始化 已经选择的token list (选中之后 再退出) 也就是初始化 fetchTokenList
+        this.props.dispatch(initSelectedListAction())
+        return;
       }
+      return;
     })
   }
 
-  onFetch = (addr) => {
-    if(!tk_db){
-        tk_db = tkSqLite.open()
-    }
-    tk_db.transaction((tx) => {
-      tx.executeSql(" select * from token ",[],(tx,results) => {
-        let len = results.rows.length
-        let selArr = []
-        for(let i = 0; i < len; i ++ ){
-          let data = results.rows.item(i)
-          if(data.tk_selected === 1){
-            selArr.push(data)
-          }         
-        }
-        this.setState({
-          selectedAssetsList: selArr
-        })
-        this.props.dispatch(refreshTokenInfoAction(addr))
-        this.props.dispatch(initSelectedListAction(selArr,addr))
-      },(error) => {
-        // console.error(error)
-        this.props.dispatch(insertToTokenAction(addr))
-      })
-    })
-  }
+
 
   componentWillReceiveProps(nextProps){
-    const { selectedAssetsList } = this.state //  willmount拉去的数据库中的数据   
-    const { selectedList,refreshEnd } = this.props.tokenManageReducer //选中或取消选中操作后得到的数据
-    if(selectedList !== nextProps.tokenManageReducer.selectedList){
-      this.setState({
-        selectedAssetsList:  nextProps.tokenManageReducer.selectedList
-      })
-    }
+ 
+    const { fetchTokenList } = this.props.tokenManageReducer
 
+    //刷新
+    const { refreshEnd} = nextProps.tokenManageReducer
 
-    if(refreshEnd !== nextProps.tokenManageReducer.refreshEnd && nextProps.tokenManageReducer.refreshEnd){
-      this.setState({
-        isRefreshing: false
-      })
-
-      web3.eth.getBalance(`0x${this.state.curAddr}`).then((res,rej)=>{
+    switch(refreshEnd){
+      case 'start':
         this.setState({
-          etzBalance: web3.utils.fromWei(res,'ether')
+          isRefreshing: true
         })
+        break
+      case 'suc':
+        this.setState({
+          isRefreshing: false
+        })
+        break
+      case 'fail':
+        this.setState({
+          isRefreshing: false
+        })
+        default:
+          break
+    }
+    
+    //删除了当前账号  需要更新reducers里的当前账号
+    const { globalAccountsList, currentAccount } = nextProps.accountManageReducer
+    if(this.props.accountManageReducer.deleteCurrentAccount !== nextProps.accountManageReducer.deleteCurrentAccount && nextProps.accountManageReducer.deleteCurrentAccount){
+      this.props.dispatch(globalCurrentAccountInfoAction(globalAccountsList[0]))
+      this.props.dispatch(refreshTokenAction(globalAccountsList[0].address,fetchTokenList))
+      this.setState({
+        navTitle: globalAccountsList[0].account_name
       })
-
+    }
+    //切换账号
+    if(this.props.accountManageReducer.currentAccount.account_name !== currentAccount.account_name){
+      this.setState({
+        navTitle: currentAccount.account_name,
+        curAddr: currentAccount.address
+      })
+      return;
     }
 
-    if(this.props.accountManageReducer.accountInfo !== nextProps.accountManageReducer.accountInfo){
-      toSplash()
+    //交易状态
+
+    const { txEtzStatus,txEtzHash, txErrorMsg,txErrorOrder,txStateMark } = nextProps.tradingManageReducer
+    if(this.props.tradingManageReducer.txEtzStatus !== txEtzStatus){
+      if(txEtzStatus === 1){
+          //更新轉賬狀態
+         Alert.alert(I18n.t('send_successful'))
+         this.updatePending(1,txEtzHash,txStateMark)
+         this.props.dispatch(resetTxStatusAction())
+         this.props.dispatch(refreshTokenAction(this.state.curAddr,fetchTokenList))
+         return
+       }else if(txEtzStatus === 0){
+      
+          if(Platform.OS === 'ios'){
+            setTimeout(() => {
+              Alert.alert(txErrorMsg)
+            },2000)
+          }else{
+            Alert.alert(txErrorMsg)
+          }
+          if(txErrorOrder === 1){
+            this.updatePending(0,txEtzHash,txStateMark)
+          }else if(txErrorOrder === 0){
+            this.deletePending(txStateMark)
+          }
+          this.props.dispatch(resetTxStatusAction())
+          return
+       }
+      
     }
   }
 
+  async deletePending(mark){
+    let delRes = await accountDB.deleteAccount({
+      sql: 'delete from trading where tx_random =  ?',
+      d_id: [mark],
+    })
+    console.log('删除结果',delRes)
 
- 
-  toAssetsDetail = (title,balance,token) => {
+    if(delRes === 'success'){
+      this.props.dispatch(updateTxListAction(true))
+    }
+
+  }
+
+  async updatePending(status,hash,mark){
+    let block = '',
+        time = '';
+    if(hash.length === 66){
+      let tx = await web3.eth.getTransaction(hash)
+      let txBlock  = await web3.eth.getBlock(tx.blockNumber)
+      block = txBlock.number
+      time = txBlock.timestamp
+    }else{
+      time =  `${Date.parse(new Date())/1000}`
+    }
+    let updateRes = await accountDB.updateTable({
+      sql: 'update trading set tx_result = ?,tx_time = ?,tx_hash = ?,tx_block_number = ? where tx_random = ? ',
+      parame: [status,time,hash,block,mark]
+    })
+
+    if(updateRes === 'success'){
+      console.log('updatePending成功')
+      //交易成功 更新tx list状态
+      this.props.dispatch(updateTxListAction(true))
+
+    }else{
+      if(updateRes === 'fail'){
+         console.log('updatePending失败')
+      }
+    }
+    
+  }
+
+  componentWillUnmount(){
+    this.onCloseDrawer()
+    AppState.removeEventListener('change', this._handleAppStateChange);
+
+  }
+  _handleAppStateChange = (nextAppState) => {
+    if(nextAppState === 'active'){
+      console.log('开始刷新',nextAppState)
+      this.getAllAccounts()
+    }
+  }
+  toAssetsDetail = (title,balance,token,deci) => {
     this.props.navigator.push({
-      screen: 'asset_detail_list',
+      screen: 'tx_record_list',
       title,
-      navigatorStyle: MainThemeNavColor,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+      // navigatorStyle: MainThemeNavColor,
+      navigatorStyle: Object.assign({},DetailNavigatorStyle,{
+          navBarHidden: true,
+          // navBarTextColor:'#fff',
+          // navBarBackgroundColor:'#144396',
+          statusBarColor:'#144396',
+          statusBarTextColorScheme:'light'
+      }),
       passProps:{
         etzBalance: balance,
         etz2rmb: 0,
         curToken: token,
+        currencySymbol: this.state.currencySymbol,
+        curDecimals: deci
       }
     })
   }
@@ -149,12 +336,17 @@ class Assets extends Component{
     this.props.navigator.push({
       screen: 'scan_qr_code',
       title:I18n.t('scan'),
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       navigatorStyle: Object.assign({},DetailNavigatorStyle,{
         navBarTextColor:'#fff',
         navBarBackgroundColor:'#000',
         statusBarColor:'#000',
         statusBarTextColorScheme:'light',
       }),
+      passProps: {
+        fromHome:true
+      }
     })
   }
 
@@ -162,9 +354,15 @@ class Assets extends Component{
     this.props.navigator.push({
       screen: 'on_payment',
       title:I18n.t('send'),
-      navigatorStyle: DetailNavigatorStyle,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+      navigatorStyle: Object.assign({},DetailNavigatorStyle,{
+        navBarHidden: true
+      }),
       passProps:{
-        curToken: 'ETZ'
+        curToken: 'ETZ',
+        currencySymbol: this.state.currencySymbol,
+        curDecimals: 0
       }
     })
   }
@@ -172,6 +370,8 @@ class Assets extends Component{
     this.props.navigator.push({
       screen: 'on_receive',
       title:I18n.t('receive'),
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       navigatorStyle: DetailNavigatorStyle,
     })
   }
@@ -179,6 +379,8 @@ class Assets extends Component{
     this.props.navigator.push({
       screen: 'trading_record',
       title:I18n.t('tx_records'),
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       navigatorStyle: Object.assign({},MainThemeNavColor,{navBarNoBorder:true}),
       // navigatorButtons: {
       //   rightButtons: [
@@ -196,6 +398,8 @@ class Assets extends Component{
       screen: 'add_assets',
       title:I18n.t('add_assets'),
       navigatorStyle: DetailNavigatorStyle,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       // navigatorButtons: {
       //   rightButtons: [
       //     {
@@ -209,12 +413,11 @@ class Assets extends Component{
 
   onDrawerCloseStart = () => {
     switchDrawer(false)
-    // this.props.dispatch(onSwitchDrawerAction(0))
   }
   onDrawerOpenStart = () => {
     switchDrawer(true)
-    // this.props.dispatch(onSwitchDrawerAction(1))
   }
+  
   onCloseDrawer = () => {
     this._drawer.close()
   }
@@ -223,6 +426,8 @@ class Assets extends Component{
       screen:'msg_center_list',
       title:I18n.t('msg_center'),
       navigatorStyle: DetailNavigatorStyle,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       navigatorButtons: {
         rightButtons: [
           {
@@ -239,29 +444,65 @@ class Assets extends Component{
   }
 
   onRefresh = () => {
+    const { fetchTokenList } = this.props.tokenManageReducer
     this.setState({
         isRefreshing: true
     })
-    this.props.dispatch(refreshTokenInfoAction(this.state.curAddr))
+    //这里的下拉刷新  更新etz和代币的余额
+    this.props.dispatch(refreshTokenAction(this.state.curAddr,fetchTokenList))
+  }
+
+  onBindPhone = () => {
+    this.props.navigator.push({
+      screen: 'bind_phone',
+      title:I18n.t('bind_phone'),
+      navigatorStyle: DetailNavigatorStyle,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+    })
+    // this.props.navigator.push({
+    //         screen: 'recom_prize',
+    //         title:I18n.t('recom_prize'),
+    //         navigatorStyle: DetailNavigatorStyle,
+    //         backButtonTitle:I18n.t('back'),
+    //         backButtonHidden:false,
+    //         passProps: {
+    //           inviteCode: '12345678'
+    //         }
+    //     })
+  }
+
+  onReceiveCandy = () => {
+    this.props.navigator.push({
+      screen: 'receive_candy',
+      title:'领取',
+      navigatorStyle: DetailNavigatorStyle,
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+    })
   }
   render(){
-    const { selectedAssetsList,etzBalance, isRefreshing} = this.state
+    const { selectedAssetsList, isRefreshing, currencySymbol} = this.state
     
+    const { currentAccount, globalAccountsList } = this.props.accountManageReducer
+    const { fetchTokenList,etzBalance } = this.props.tokenManageReducer
+    // console.log('etzBalance  render===',etzBalance)
+    // console.log('deviceWidth  render===',deviceWidth)
+    // console.log('deviceHeight  render===',deviceHeight)
 
     return(
-      <View style={{backgroundColor:'#F5F7FB',flex:1}}>
+      <View style={styles.containerView}>
         {
           Platform.OS === 'ios' ? 
-          <View style={{height: scaleSize(40),backgroundColor:'#144396'}}/>
+          <View style={styles.stateBar}/>
           : null
-        }
+        }     
         <Drawer
           ref={(ref) => this._drawer = ref}
           type="overlay"
           openDrawerOffset={0.4}
           side={'right'}
           tapToClose={true}
-          ref={(ref) => this._drawer = ref}
           content={<SwitchWallet thisPorps={this} onCloseSwitchDrawer={this.onCloseDrawer}/>}
           onCloseStart={this.onDrawerCloseStart}
           onOpenStart={this.onDrawerOpenStart}
@@ -280,26 +521,31 @@ class Assets extends Component{
               />
             }
           >
-            <View style={[styles.navbarStyle,pubS.center,{paddingLeft: scaleSize(24),paddingRight: scaleSize(24)}]}>
+            <View style={[styles.navbarStyle,pubS.rowCenterJus,{paddingLeft: scaleSize(24),paddingRight: scaleSize(24)}]}>
               {
                 // <TouchableOpacity activeOpacity={.6} onPress={this.onLeftDrawer}>
                 //   <Image source={require('../../images/xhdpi/nav_ico_home_message_def.png')}style={styles.navImgStyle}/>
                 // </TouchableOpacity>
                 
+              // <TouchableOpacity activeOpacity={.6} onPress={this.onReceiveCandy} style={{width: scaleSize(160),height: scaleSize(87),justifyContent:'center'}}> 
+              //   <Text style={pubS.font26_1}>领取</Text>
+              // </TouchableOpacity>
               }
-              <Text style={pubS.font30_1}>{this.state.navTitle}</Text>
-              <TouchableOpacity activeOpacity={.6} onPress={this.onRightDrawer} style={styles.drawerStyle}>
+              <View style={{width: scaleSize(160),height: scaleSize(87),justifyContent:'center'}}>
+              </View>
+              <Text style={[pubS.font30_1,{}]}>{this.state.navTitle}</Text>
+              <TouchableOpacity activeOpacity={.6} onPress={() => this.onRightDrawer()} style={[styles.drawerStyle,{}]}>
                 <Image source={require('../../images/xhdpi/nav_ico_home_more_def.png')} style={styles.navImgStyle}/>
               </TouchableOpacity>
             </View>
             <View>
               <View style={[styles.assetsTotalView,pubS.center,{height: Platform.OS === 'ios' ? scaleSize(260) : scaleSize(300)}]}>
-                  <Text style={pubS.font72_1}>≈0</Text>
-                  <Text style={pubS.font26_3}>{I18n.t('total_assets')}(¥)</Text>
+                  <Text style={pubS.font72_1}>{splitDecimal(etzBalance)}</Text>
+                  <Text style={pubS.font26_3}>{I18n.t('total_assets')}</Text>
               </View>
 
               <View style={[styles.optionView,pubS.center]}>
-                  <View style={[pubS.rowCenterJus,{width: scaleSize(650)}]}>
+                  <View style={[pubS.rowCenterJus,styles.listItemBox]}>
                     <TouchableOpacity activeOpacity={.7} onPress={this.onScan} style={[styles.optionItem]}>
                       <Image source={require('../../images/xhdpi/btn_ico_home_scan_def.png')} style={styles.itemImageStyle}/>
                       <Text style={[pubS.font24_2,]}>{I18n.t('scan')}</Text>
@@ -323,21 +569,27 @@ class Assets extends Component{
               shortName={etzTitle}
               fullName={'EtherZero'}
               coinNumber={splitDecimal(etzBalance)}
-              price2rmb={0}
-              onPressItem={() => this.toAssetsDetail(etzTitle,splitDecimal(etzBalance),'ETZ')}
+              //price2rmb={0}
+              symbol={this.state.currencySymbol}
+              onPressItem={() => this.toAssetsDetail(etzTitle,splitDecimal(etzBalance),'ETZ',0)}
             />
             {
-              selectedAssetsList.map((res,index) => {
-                return(
-                  <AssetsItem
-                    key={index}
-                    shortName={res.tk_symbol}
-                    fullName={res.tk_name}
-                    coinNumber={splitDecimal(res.tk_number)}
-                    price2rmb={0}
-                    onPressItem={() => this.toAssetsDetail(res.tk_symbol,splitDecimal(res.tk_number),res.tk_symbol)}
-                  />
-                )
+              fetchTokenList.map((res,index) => {
+                // console.log('res.account_addr===',res.account_addr)
+                // console.log('currentAccount.address=',currentAccount.address)
+                if(res.tk_selected === 1 && (res.account_addr === currentAccount.address)){
+                  return(
+                    <AssetsItem
+                      key={index}
+                      shortName={res.tk_symbol}
+                      fullName={res.tk_name}
+                      coinNumber={scientificToNumber(splitDecimal(res.tk_number)) }
+                      //price2rmb={0}
+                      symbol={this.state.currencySymbol}
+                      onPressItem={() => this.toAssetsDetail(res.tk_symbol,splitDecimal(res.tk_number),res.tk_symbol,res.tk_decimals)}
+                    />
+                  )
+                }
               })
             }
             <TouchableOpacity style={[styles.whStyle,styles.addBtnStyle,pubS.center]} activeOpacity={.7} onPress={this.addAssetsBtn}>
@@ -358,29 +610,49 @@ class AssetsItem extends Component {
     return(
       <TouchableOpacity style={[styles.listItemView,styles.whStyle]} activeOpacity={.7} onPress={onPressItem}>
         <Image source={require('../../images/xhdpi/etz_logo.png')} style={pubS.logoStyle}/>
-        <View style={[styles.listItemTextView]}>
-          <View style={pubS.rowCenterJus}>
+        <View style={[styles.listItemTextView,pubS.rowCenterJus]}>
+          <View>
             <Text style={pubS.font36_2}>{shortName}</Text>
+            <Text style={[pubS.font24_2]}>{fullName}</Text>
+          </View>
+          <View>
             <Text style={pubS.font36_2}>{coinNumber}</Text> 
           </View>
-          <View style={pubS.rowCenterJus}>
-            <Text style={pubS.font24_2}>{fullName}</Text>
-            <Text style={pubS.font24_2}>{`≈ ¥${price2rmb}`}</Text>
-          </View>
+            {
+          //<View style={pubS.rowCenterJus}>
+              //<Text style={pubS.font24_2}>{`≈ ${this.props.symbol}${price2rmb}`}</Text>
+          //</View>
+            }
         </View>
       </TouchableOpacity>
     )
   }
 }
 const styles = StyleSheet.create({
+  containerView:{
+    backgroundColor:'#F5F7FB',
+    flex:1,
+  },
+  stateBar:{ 
+    ...ifIphoneX(
+      {
+        height: 44,
+        backgroundColor:'#144396'
+      },
+      {
+        height: scaleSize(46),
+        backgroundColor:'#144396'
+      }
+    )
+  },
   drawerStyle:{
     // borderColor:'#fff',
     // borderWidth:1,
     height: scaleSize(87),
     width: scaleSize(160),
-    position:"absolute",
-    top: 0,
-    right:scaleSize(24),
+    // position:"absolute",
+    // top: 0,
+    // right:scaleSize(24),
     alignItems:'flex-end',
     justifyContent:'center'
   },
@@ -400,22 +672,89 @@ const styles = StyleSheet.create({
     marginTop: scaleSize(20),
     alignSelf:'center',
   },
+  logoStyle:{
+    ...ifIphoneX({
+      width: 27,
+      height:27,
+      marginTop: 16
+    },
+    {
+      width: scaleSize(44),
+      height:scaleSize(44),
+      marginTop: scaleSize(22)
+    },
+    {
+      width: scaleSize(44),
+      height:scaleSize(44),
+      marginTop: scaleSize(22)
+    }
+    )
+  },
   whStyle: {
-    height: scaleSize(120),
-    width: scaleSize(702),
+    ...ifIphoneX(
+      {
+        width:scaleSize(580),
+        height: scaleSize(120)
+      },
+      {
+        width:scaleSize(702),
+        height: scaleSize(120)
+      },
+      {
+        width:scaleSize(702),
+        height: scaleSize(120)
+      }
+    )
+
   },
   listItemTextView:{
-    width: scaleSize(618),
-    marginLeft:scaleSize(18),
-    paddingTop: scaleSize(15),
-    paddingBottom: scaleSize(22),
-    // borderColor:'red',
-    // borderWidth:1,
+    ...ifIphoneX(
+      {
+        width: 293,
+        marginLeft:scaleSize(18),
+        // paddingTop: scaleSize(15),
+        // paddingBottom: scaleSize(22),
+      },
+      {
+        width: scaleSize(618),
+        marginLeft:scaleSize(18),
+        // paddingTop: scaleSize(15),
+        // paddingBottom: scaleSize(22),
+      },
+      {
+        width: scaleSize(618),
+        marginLeft:scaleSize(18),
+        // paddingTop: scaleSize(15),
+        // paddingBottom: scaleSize(22),
+        // borderColor:'red',
+        // borderWidth:1,
+      }
+    )
+    
+    
+  },
+  listItemBox:{
+    ...ifIphoneX(
+      {
+        width:345,
+        alignSelf:'center',
+        flex:1,
+      },
+      {
+        width:scaleSize(650),
+        alignSelf:'center',
+        flex:1,
+      },
+      {
+        width:scaleSize(650),
+        alignSelf:'center',
+        flex:1,
+      }
+  )
   },
   listItemView:{
     backgroundColor:'#fff',
-    paddingLeft: scaleSize(22),
-    paddingRight: scaleSize(22),
+    ...ifIphoneX({marginLeft:30,marginRight:30},{paddingLeft: scaleSize(22),paddingRight: scaleSize(22)},{paddingLeft: scaleSize(22),paddingRight: scaleSize(22)}),
     justifyContent:'center',
     flexDirection:'row',
     borderRadius: 4,
@@ -439,6 +778,10 @@ const styles = StyleSheet.create({
     backgroundColor:'#fff',
     // borderColor:'red',
     // borderWidth:1,
+    ...ifIphoneX({
+      width: 375,
+      alignSelf:'center'
+    })
   },
   assetsTotalView: {
     
@@ -450,6 +793,7 @@ const styles = StyleSheet.create({
 export default connect(
   state => ({
     accountManageReducer: state.accountManageReducer,
-    tokenManageReducer: state.tokenManageReducer
+    tokenManageReducer: state.tokenManageReducer,
+    tradingManageReducer: state.tradingManageReducer
   })
 )(Assets)
