@@ -6,11 +6,15 @@ import {
   TouchableOpacity,
   StyleSheet,
   ScrollView,
+  BackHandler,
+  StatusBar,
+  Platform
 } from 'react-native'
 
 import { pubS,DetailNavigatorStyle} from '../../styles/'
-import { setScaleText, scaleSize } from '../../utils/adapter'
+import { setScaleText, scaleSize,ifIphoneX } from '../../utils/adapter'
 import I18n from 'react-native-i18n'
+import { onExitApp } from '../../utils/exitApp'
 class Login extends Component{
   constructor(props){
     super(props)
@@ -18,16 +22,24 @@ class Login extends Component{
 
     }
   }
-  componentDidMount(){
-    
+  componentWillMount(){
+    BackHandler.addEventListener('hardwareBackPress',this.onBack)
+  }
+  componentWillUnmount () {
+    BackHandler.removeEventListener('hardwareBackPress',this.onBack)
+  }
+  onBack = () => {
+    onExitApp()
   }
   createAccoumt = () => {
 
     this.props.navigator.push({
       screen: 'create_account',
       title:I18n.t('create'),
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
       navigatorStyle: Object.assign({},DetailNavigatorStyle,{navBarNoBorder:true}),
-      passProps: { from: 'login_create'}
+      passProps: { fromLogin: 'login'}
     })
   }
 
@@ -36,7 +48,10 @@ class Login extends Component{
     this.props.navigator.push({
       screen: 'import_account',
       title:I18n.t('import'),
-      navigatorStyle: DetailNavigatorStyle
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+      navigatorStyle: DetailNavigatorStyle,
+      passProps: { fromLogin: 'login'}
     })
   }
   // termsOfService = () => {
@@ -50,6 +65,11 @@ class Login extends Component{
 
     return(
       <View style={pubS.container}>
+        {
+          Platform.OS === 'ios' ?
+            <StatusBar backgroundColor="#FFFFFF"  barStyle="light-content"  animated={true}/>
+          : null
+        }
         <Image source={require('../../images/xhdpi/bg_signin.png')} style={[pubS.fullWH,styles.bgStyle]}/>
         <Image source={require('../../images/xhdpi/logo.png')} style={styles.logoStyle}/>
         <View style={styles.btnContainer}>
@@ -77,20 +97,62 @@ const styles = StyleSheet.create({
     // left:0,
   },
   logoStyle: {
-    position:'absolute',
-    top: scaleSize(250),
-    left: scaleSize(292),
-    width: scaleSize(166),
-    height: scaleSize(242),
+    ...ifIphoneX(
+      {
+        position:'absolute',
+        top: scaleSize(250),
+        left: 135 ,
+        width: scaleSize(166),
+        height: scaleSize(242),
+      },
+      {
+        position:'absolute',
+        top: scaleSize(250),
+        left: scaleSize(292),
+        width: scaleSize(166),
+        height: scaleSize(242),
+      },
+      {
+        position:'absolute',
+        top: scaleSize(250),
+        left: scaleSize(292),
+        width: scaleSize(166),
+        height: scaleSize(242),
+      }
+    )
+    
   },
   btnContainer:{
-    justifyContent:'center',
-    position:'absolute',
-    top: scaleSize(748),
-    width: scaleSize(750),
-    alignItems:'center',
-    // borderColor:'red',
-    // borderWidth:1,
+    ...ifIphoneX(
+      {
+        justifyContent:'center',
+        position:'absolute',
+        top: scaleSize(748),
+        width: 376,
+        alignItems:'center',
+        // borderColor:'red',
+        // borderWidth:1,
+      },
+      {
+        justifyContent:'center',
+        position:'absolute',
+        top: scaleSize(748),
+        width: scaleSize(750),
+        alignItems:'center',
+        // borderColor:'red',
+        // borderWidth:1,
+      },
+      {
+        justifyContent:'center',
+        position:'absolute',
+        top: scaleSize(748),
+        width: scaleSize(750),
+        alignItems:'center',
+        // borderColor:'red',
+        // borderWidth:1,
+      }
+    )
+    
   },
   btnStyle: {
     height: scaleSize(94),
