@@ -3,19 +3,67 @@ import {
   View,
   Text,
   WebView,
+  StyleSheet,
+  StatusBar,
+  Platform,
+  ActivityIndicator
 } from 'react-native'
 
 import { pubS,DetailNavigatorStyle } from '../../styles/'
-import { setScaleText, scaleSize } from '../../utils/adapter'
+import { setScaleText, scaleSize,ifIphoneX } from '../../utils/adapter'
 
 class HelpCenter extends Component{
+  renderLoading = () => {
+		return(
+			<View style={{flex:1,alignSelf:'center',marginTop: '50%'}}>
+	           <ActivityIndicator  
+	              color={'#144396'}
+	              indeterminate={true}
+	              size={'large'}
+	            />
+	        </View> 
+		)
+	}
   render(){
     return(
+      <View style={{flex:1}}>
+        {
+          Platform.OS === 'ios' ?
+          <StatusBar backgroundColor={"#000000"}  barStyle={"dark-content"}  animated={true} />
+          : null
+        }
         <WebView  
-          source={{uri:'https://mp.weixin.qq.com/s/8RQ8GOi0d2z03vYcTjsCrw'}}
-          style={{flex:1,alignSelf:'center',width: scaleSize(750)}}
+          source={{uri:'https://mp.weixin.qq.com/s/vj9ZTq3uNBeHL6W5PuhEnQ'}}
+          style={styles.webViewStyle}
+          startInLoadingState={true}
+          renderLoading={this.renderLoading }
+          domStorageEnabled={true}//开启dom存贮
+          javaScriptEnabled={true}//开启js
         />
+      </View>
     )
   }
 }
+
+const styles = StyleSheet.create({
+   webViewStyle:{
+     ...ifIphoneX(
+       {
+         width:375,
+         flex:1,
+         alignSelf:'center',
+       },
+       {
+        width: scaleSize(750),
+        flex:1,
+        alignSelf:'center',
+       },
+       {
+        width: scaleSize(750),
+        flex:1,
+        alignSelf:'center',
+       }
+     )
+   }
+})
 export default HelpCenter
