@@ -11,23 +11,29 @@ import {
   BackHandler,
   Clipboard,
   Share,
+  Button,
+  StatusBar,
+  Platform
 } from 'react-native'
 import { Navigation } from 'react-native-navigation'
 import { pubS,DetailNavigatorStyle } from '../../../styles/'
-import { setScaleText, scaleSize } from '../../../utils/adapter'
+import { setScaleText, scaleSize, ifIphoneX } from '../../../utils/adapter'
 import { sliceAddress } from '../../../utils/splitNumber'
 import { Btn,Loading, } from '../../../components/'
 import Modal from 'react-native-modal'
 import { connect } from 'react-redux'
-import UserSQLite from '../../../utils/accountDB'
 
-import { deleteAccountAction,resetDeleteStatusAction,updateBackupStatusAction,passAccountsInfoAction } from '../../../actions/accountManageAction'
+
+import { deleteAccountAction,resetDeleteStatusAction,updateBackupStatusAction,passPropsAction } from '../../../actions/accountManageAction'
 const Wallet = require('ethereumjs-wallet')
-const sqLite = new UserSQLite()
-let db
+
 import { toLogin } from '../../../root'
 import I18n from 'react-native-i18n'
-import Toast from 'react-native-toast'
+import Toast from 'react-native-root-toast'
+
+import accountDB from '../../../db/account_db'
+
+import { fromV3 } from '../../../utils/fromV3'
 class BackUpAccount extends Component{
   constructor(props){
     super(props)
@@ -36,72 +42,38 @@ class BackUpAccount extends Component{
       pKeyVisible: false,
       psdVal: '',
       privKey: '',
-      privBackuped: false,
       backupMnemonic: false,
-      mncBackuped: false,
       isDelAccount: false,
       dVisible: false,
       keyStore: {},
       loadingText: '',
       visible: false,
-      time1:'',
-      time2:'',
+      localMnemonic: '',
+      currentList: {},
+      pasModalVisible: -1,//0  私钥 1 删除  2 keystore 3助记词
     }
     this.props.navigator.setOnNavigatorEvent(this.onNavigatorEvent.bind(this))
   }
   componentWillMount(){
-    this.props.dispatch(resetDeleteStatusAction())
-    if(!db){  
-        db = sqLite.open();  
-    }  
-    db.transaction((tx)=>{  
-      tx.executeSql("select * from account where address= ? ", [this.props.address],(tx,results)=>{  
-        let res = results.rows.item(0)
-        let keyStore = {
-          "version":res.version,
-          "id":res.kid,
-          "address":res.address,
-          "crypto":{
-            "ciphertext":res.ciphertext,
-            "cipherparams":{
-                "iv":res.iv
-            },
-            "cipher":res.cipher,
-            "kdf":res.kdf,
-            "kdfparams":{
-              "dklen":res.dklen,
-              "salt":res.salt,
-              "n":res.n,
-              "r":res.r,
-              "p":res.p
-            },
-            "mac":res.mac
-          }
-        }
-        this.setState({
-          keyStore
-        })
-        if(res.backup_status === 1){
-          this.setState({
-            privBackuped: true
-          })
-        }
-        if(!res.mnemonic){
-          this.setState({
-            mncBackuped: true
-          })
-        }
+    const { currentAccount, globalAccountsList } = this.props.accountManageReducer
 
-      });  
-    },(error)=>{
-      console.error(error)
-    }); 
+    this.props.dispatch(resetDeleteStatusAction())
+    
+    globalAccountsList.map((list,index) => {
+      if(list.address === this.props.address){
+        this.setState({
+          localMnemonic: list.mnemonic,
+          currentList: list
+        })
+        
+       
+        let keys = this.getKeys(list)
+
+        this.setState({ keyStore:keys })
+      }
+    })
   }
 
-
-  compennetDidUnmount(){  
-    sqLite.close()  
-  } 
 
   componentWillReceiveProps(nextProps){
     if(this.props.accountManageReducer.deleteSuc !== nextProps.accountManageReducer.deleteSuc && nextProps.accountManageReducer.deleteSuc){
@@ -109,59 +81,105 @@ class BackUpAccount extends Component{
         visible: false,
         loadingText: ''
       })
-      Toast.showLongBottom(I18n.t('delete_successfully'))
+      let t = Toast.show(I18n.t('delete_successfully'))
+      setTimeout(() => {
+        Toast.hide(t)
+      },1000)
+
       if(this.props.accountsNumber === 1){
         setTimeout(() => {
           toLogin()
         },1000)
-        sqLite.dropTable()
-        sqLite.deleteData()  
-      }else{
-        //删除其他账号后 更新accountInfo信息  如果删除的是当前账号  更新accountInfo后还需要将另外的一条信息的is_selected=1
-        this.setState({
-          visible: false,
-          loadingText: ''
+        accountDB.dropTable({
+          sql: 'drop table account'
         })
-        this.props.dispatch(passAccountsInfoAction())
+
+      }else{
         setTimeout(() => {
           this.props.navigator.pop()
         },1000)
       }
     }
 
-    if(this.props.accountManageReducer.delMnemonicSuc !== nextProps.accountManageReducer.delMnemonicSuc && nextProps.accountManageReducer.delMnemonicSuc){
-      this.setState({
-        mncBackuped: true
-      })
-    }
+
+    //修改密码成功后  keystore需要使用reducer中的
+    const { globalAccountsList,copyKeystoreSuc } = nextProps.accountManageReducer
+
+    console.log('修改密码成功后  keystore需要使用reducer中的',globalAccountsList)
+    globalAccountsList.map((list,index) => {
+      if(list.address === this.props.address){
+        this.setState({
+          currentList: list
+        })
+        
+        let keys = this.getKeys(list)
+        this.setState({ keyStore:keys })
+      }
+    })
+
 
   }
-  onNavigatorEvent(event){
-    if (event.type == 'NavBarButtonPress') {
-      switch(event.id){
-        case 'save_back_up_info':
-          alert('save')
-          break
-        // case 'backPress':
-        //   this.props.navigator.resetTo({
-        //     screen: 'account_manage',
-        //     title:'Manage wallets',
-        //     overrideBackPress: true,
-        //     navigatorStyle: DetailNavigatorStyle,
-        // })
-          break
-        default:
-          break
+
+  getKeys(list){
+    let keys = {
+      "version":list.version,
+      "id":list.kid,
+      "address":list.address,
+      "crypto":{
+        "ciphertext":list.ciphertext,
+        "cipherparams":{
+            "iv":list.iv
+        },
+        "cipher":list.cipher,
+        "kdf":list.kdf,
+        "kdfparams":{
+          "dklen":list.dklen,
+          "salt":list.salt,
+          "n":list.n,
+          "r":list.r,
+          "p":list.p
+        },
+        "mac":list.mac
       }
     }
+    return keys
   }
 
+  onNavigatorEvent(event){
+
+    if (event.type == 'NavBarButtonPress') {
+      if(event.id === 'save_change'){
+        this.saveUserName()
+      }
+    }
+     switch (event.id) {
+      case 'willAppear':
+        this.backHandler = BackHandler.addEventListener('hardwareBackPress', this.handleBackPress);
+        break;
+      case 'willDisappear':
+        this.backPressed = 0;
+        this.backHandler.remove();
+        break;
+      default:
+        break
+     }
+  }
+  handleBackPress = () => {
+    if (this.backPressed && this.backPressed > 0) {
+      this.props.navigator.popToRoot({ animated: false })
+      return false
+    }
+    this.backPressed = 1
+  }
+  saveUserName = () => {
+
+  }
   deleteAccount = () => {
     this.setState({
       iptPsdVisible: true,
-      isDelAccount: true
+      isDelAccount: true,
+      pasModalVisible: 1
     })
-
   }
 
   onPressCancel = () => {
@@ -170,6 +188,7 @@ class BackUpAccount extends Component{
     })
   }
   onPressConfirmDel = () => {
+    console.log('onPressConfirmDel1111')
     this.setState({
       dVisible: false,
       visible: true,
@@ -180,14 +199,16 @@ class BackUpAccount extends Component{
 
   backUpPrivBtn = () => {
     this.setState({
-      iptPsdVisible: true
+      iptPsdVisible: true,
+      pasModalVisible: 0,
     })
   }
   onHide = () => {
     this.setState({
       iptPsdVisible: false,
       psdVal: '',
-      backupMnemonic: false,
+      // backupMnemonic: false,
+      pasModalVisible: -1,
       isDelAccount: false,
       visible: false
     })
@@ -201,123 +222,175 @@ class BackUpAccount extends Component{
     this.setState({
       pKeyVisible: false,
       privKey: '',
-      privBackuped: true,
     })
-
-    this.props.dispatch(updateBackupStatusAction(this.props.address))
   }
 
   onConfirm = () => {
-    const { psdVal,backupMnemonic,keyStore,isDelAccount,  } = this.state
+    const { psdVal,backupMnemonic,keyStore,isDelAccount } = this.state
     this.setState({
       loadingText: I18n.t("unlocking"),
-      visible: true,
-      iptPsdVisible: false
-    })  
+      iptPsdVisible: false,
+    })
     setTimeout(() => {
+      this.setState({
+        visible: true,
+      })
+    },500)
+    setTimeout(async () => {
       try {
-        let d1 = new Date()
-        // this.setState({
-        //   time1: `${d1.getMinutes()}分--${d1.getSeconds()}秒`
-        // })
-        const newWallet = Wallet.fromV3(keyStore,psdVal)
-        // let d2 = new Date()
-        // this.setState({
-        //   time2: `${d2.getMinutes()}分--${d2.getSeconds()}秒`
-        // })
-        let priv = newWallet._privKey.toString('hex')
-        if(backupMnemonic){
-          this.props.navigator.push({
-            screen: 'write_mnemonic',
-            title: '',
-            navigatorStyle: DetailNavigatorStyle,
-            passProps: {
-              currentAddress: this.props.address
-            }
-          })
-        }else{
-          if(isDelAccount){
-            this.setState({
-              dVisible: true,
-            })
-          }else{
-            this.setState({
-              privKey: priv,
-              pKeyVisible: true
+          const priv = await fromV3(keyStore,psdVal)
+          this.showDiffModal(priv)
+
+          this.onHide()
+
+      } catch (err) {
+        console.log('fromV3 err',err)
+        Alert.alert(
+          I18n.t('title_error'),
+          `${err}`,
+          [
+            {text:I18n.t('ok'),onPress:() => 
+            this.setState({ 
+              psdVal: '',
+              loadingText: '',
+              visible: false,
             })
           }
-        }
-        this.onHide()
-      } catch (err) {
-        Toast.showLongBottom(I18n.t('password_is_wrong'))
-        this.setState({
-          psdVal: '',
-          visible: false,
-          loadingText: '',
-        })
+          ]
+        ) 
       }
     },1000)
   }
+  showDiffModal(priv){
+    switch(this.state.pasModalVisible){//0  私钥 1 删除  2 keystore 3助记词
+      case 0:
+        this.showPrivModal(priv)
+        break
+      case 1:
+        this.showDelModal()
+        break
+      case 2:
+        this.toExportKey()
+        break
+      case 3:
+        this.showMneModal()
+        break
+      default:
+        break
+    }
+  }
+  showPrivModal = (priv) => {
+    setTimeout(() => {
+      this.setState({
+        privKey: priv,
+        pKeyVisible: true,
+      })
+    },1000)
+  }
+  showDelModal = () => {
+    setTimeout(() => {
+      this.setState({
+        dVisible: true,
+      })
+    },1000) 
+  }
+  toExportKey = () => {
+    const { currentList, keyStore} = this.state
+    this.props.dispatch(passPropsAction({
+      currentList,
+      keyStore
+    }))
 
-
+    this.props.navigator.push({
+      screen: 'export_keystore',
+      title:I18n.t('export_keystore'),
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+      navigatorStyle: DetailNavigatorStyle,
+      passProps:{
+        curAddr:this.props.address
+      }
+    })
+  }
+  showMneModal = () => {
+    this.props.navigator.push({
+      screen: 'write_mnemonic',
+      title: '',
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+      navigatorStyle: DetailNavigatorStyle,
+      passProps: {
+        currentAddress: this.props.address,
+        localMnemonic: this.state.localMnemonic
+      }
+    })
+  }
 
 
   onCopyBtn = () => {
     Clipboard.setString(this.state.privKey)
-    Toast.showLongBottom(I18n.t('copy_successfully'))
+    let t = Toast.show(I18n.t('copy_successfully'))
+    setTimeout(() => {
+      Toast.hide(t)
+    },1000)
+    this.props.dispatch(updateBackupStatusAction(this.props.address))
   }
-  backupMnemonicBtn = () => {
+
+  backUpKeyStoreBtn = () => {
+
     this.setState({
       iptPsdVisible: true,
-      backupMnemonic: true
+      pasModalVisible: 2,
     })
-
   }
-  backUpKeyStoreBtn = () => {
-    const { keyStore } = this.state
-    let k = JSON.stringify(keyStore)
-    Share.share({
-      message: k,
-      title: I18n.t('backup_keystore_title'),
-    }, {
-      dialogTitle: I18n.t('share_your_keystore'),
+
+
+
+  modifyPsd = () => {
+    const { currentList, keyStore} = this.state
+    this.props.dispatch(passPropsAction({
+      currentList,
+      keyStore
+    }))
+    this.props.navigator.push({
+      screen: 'modify_password',
+      title:I18n.t('change_psd'),
+      backButtonTitle:I18n.t('back'),
+      backButtonHidden:false,
+      navigatorStyle: DetailNavigatorStyle,
     })
-    .then(this._showResult)
-    .catch((error) => Toast.showLongBottom(I18n.t('share_error')))
   }
-  _showResult = (result) => {
-    if (result.action === Share.sharedAction) {
-      if (result.activityType) {
-
-      } else {
-
-      }
-    } else if (result.action === Share.dismissedAction) {
-      Toast.showLongBottom(I18n.t('share_error'))
-    }
-  }
-
   render(){
-    const { iptPsdVisible,psdVal,pKeyVisible,privKey,privBackuped,mncBackuped,keyStore,dVisible } = this.state
-    const { isLoading } = this.props.accountManageReducer
+    const { iptPsdVisible,psdVal,pKeyVisible,privKey,keyStore,dVisible, visible } = this.state
+    const { isLoading,delMnemonicSuc } = this.props.accountManageReducer
     return(
       <View style={[pubS.container,{backgroundColor:'#fff',alignItems:'center'}]}>
-        <Loading loadingVisible={this.state.visible} loadingText={this.state.loadingText}/>
+        <Loading loadingVisible={visible} loadingText={this.state.loadingText}/>
+        {
+          Platform.OS === 'ios' ?
+          <StatusBar backgroundColor="#000000"  barStyle="dark-content" animated={true} />
+          : null
+        }
+        
         <Image source={require('../../../images/xhdpi/Penguin.png')} style={styles.avateStyle}/>
         <Text style={pubS.font26_5}>{sliceAddress(this.props.address,10)}</Text>
-        
-        <View style={[styles.userNameViewStyle,pubS.rowCenterJus,pubS.bottomStyle]}>
+        <View style={[styles.userNameViewStyle,pubS.rowCenterJus,pubS.bottomStyle,{marginTop: scaleSize(80),}]}>
           <Text style={pubS.font26_4}>{I18n.t('account_name')}</Text>
           <Text style={pubS.font26_4}>{this.props.userName}</Text>
         </View>
+        {
+          this.props.psdPrompt.length > 0 ? 
+          <View style={[styles.userNameViewStyle,pubS.rowCenterJus,pubS.bottomStyle]}>
+            <Text style={pubS.font26_4}>{I18n.t('pas_prompt')}</Text>
+            <Text style={pubS.font26_4}>{this.props.psdPrompt}</Text>
+          </View>
+          : null
+        }
+        <TouchableOpacity style={[styles.userNameViewStyle,pubS.rowCenterJus,pubS.bottomStyle]} onPress={this.modifyPsd} activeOpacity={.7} >
+          <Text style={pubS.font26_4}>{I18n.t('change_psd')}</Text>
+          <Image source={require('../../../images/xhdpi/btn_ico_payment_select_def.png')} style={{width:scaleSize(16),height: scaleSize(30)}}/>
+        </TouchableOpacity>
         <View style={{position:'absolute',bottom: scaleSize(40)}}>
-          <Btn
-            btnPress={ mncBackuped ? () => {return} : () => this.backupMnemonicBtn() }
-            bgColor={mncBackuped ? '#BDC0C6' : '#2B8AFF'}
-            opacity={mncBackuped ? 1 : .7}
-            btnText={I18n.t('backup_mnemonic_1')}
-            btnMarginTop={scaleSize(150)}
-          />
           <Btn
             btnPress={() => this.backUpKeyStoreBtn() }
             bgColor={'#2B8AFF'}
@@ -326,9 +399,9 @@ class BackUpAccount extends Component{
             btnMarginTop={scaleSize(20)}
           />
           <Btn
-            btnPress={privBackuped ? () => {return} : () => this.backUpPrivBtn() }
-            bgColor={privBackuped ? '#BDC0C6' : '#2B8AFF'}
-            opacity={privBackuped ? 1 : .7}
+            btnPress={ () => this.backUpPrivBtn() }
+            bgColor={'#2B8AFF'}
+            opacity={.7}
             btnText={I18n.t('backup_private_key')}
             btnMarginTop={scaleSize(20)}
           />
@@ -356,6 +429,7 @@ class BackUpAccount extends Component{
               textAlignVertical={'center'}
               secureTextEntry={true}
               style={styles.textIptStyle}
+              autoFocus={true}
             />
             <View style={[pubS.rowCenter,pubS.topBorderStyle,{height: scaleSize(88),marginTop: scaleSize(25),width: '100%'}]}>
               <TouchableOpacity activeOpacity={.7} onPress={this.onHide} style={[pubS.center,styles.modalBtnStyle]}>
@@ -367,19 +441,18 @@ class BackUpAccount extends Component{
             </View>
           </View>
         </Modal>
-
         <Modal
           isVisible={pKeyVisible}
           backdropOpacity={.8}
         >
           <View style={styles.pkViewStyle}>
-            <View style={[{height: scaleSize(90),backgroundColor:'#2B8AFF',width: '100%',borderTopLeftRadius: scaleSize(10),borderTopRightRadius: scaleSize(10),},pubS.center]}>
+            <View style={[styles.privViewStyle,pubS.center]}>
               <Text style={[pubS.font36_4,{fontWeight: 'bold'}]}>{I18n.t('backup_private_key')}</Text>
               <TouchableOpacity activeOpacity={.7} onPress={this.onPKeyHide} style={styles.iconStyle}>
                 <Image source={require('../../../images/xhdpi/btn_ico_collectionnobackup_close_def.png')} style={{height: scaleSize(30),width: scaleSize(30)}}/>
               </TouchableOpacity>
             </View>
-            <View style={{backgroundColor:'#FFE186',paddingLeft: scaleSize(28),paddingRight: scaleSize(28),paddingTop: scaleSize(13),paddingBottom: scaleSize(13)}}>
+            <View style={styles.privModalView}>
               <Text style={pubS.font22_1}>{I18n.t('private_key_modal')}</Text>
             </View>
             <View style={[styles.pkStyle,pubS.center]}>
@@ -404,12 +477,28 @@ class BackUpAccount extends Component{
             </View>
           </View>
         </Modal>
+        
 
       </View>
     )
   }
 }
 const styles = StyleSheet.create({
+  privModalView: {
+    backgroundColor:'#FFE186',
+    paddingLeft: scaleSize(28),
+    paddingRight: scaleSize(28),
+    paddingTop: scaleSize(13),
+    paddingBottom: scaleSize(13)
+  },
+  privViewStyle:{
+    height: scaleSize(90),
+    backgroundColor:'#2B8AFF',
+    width: '100%',
+    borderTopLeftRadius: scaleSize(10),
+    borderTopRightRadius: scaleSize(10)
+  },
+
   btnViewStyle: {
     position: 'absolute',
     bottom:0,
@@ -492,9 +581,22 @@ const styles = StyleSheet.create({
     borderRadius: scaleSize(26),
   },
   userNameViewStyle:{
-    height:scaleSize(100),
-    width: scaleSize(680),
-    marginTop: scaleSize(80),
+    ...ifIphoneX(
+      {
+        height:scaleSize(100),
+        width: 345,
+        
+      },
+      {
+        height:scaleSize(100),
+        width: scaleSize(680),
+      },
+      {
+        height:scaleSize(100),
+        width: scaleSize(680),
+      }
+    )
+
   },
   avateStyle:{
     width: scaleSize(112),
