@@ -1,3 +1,4 @@
-import './global.js'
+// import './global.js'
+import './shim'
 import './app/utils/globalWeb3.js'
 import App from './app/root'
