@@ -1,5 +1,5 @@
-const extraNodeModules = require('node-libs-browser');
+// const extraNodeModules = require('node-libs-browser');
 
-module.exports = {
-  extraNodeModules,
-};
+// module.exports = {
+//   extraNodeModules,
+// };
